branch: introduce --set-upstream-to
The existing --set-uptream option can cause confusion, as it uses the
usual branch convention of assuming a starting point of HEAD if none
is specified, causing

    git branch --set-upstream origin/master

to create a new local branch 'origin/master' that tracks the current
branch. As --set-upstream already exists, we can't simply change its
behaviour. To work around this, introduce --set-upstream-to which
accepts a compulsory argument indicating what the new upstream branch
should be and one optinal argument indicating which branch to change,
defaulting to HEAD.

The new options allows us to type

    git branch --set-upstream-to origin/master

to set the current branch's upstream to be origin's master.

Signed-off-by: Carlos Martín Nieto <cmn@elego.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6c7bae0ee62925df218f1de2f2610845337ddfb5 [formerly 222e79ebc41a8f237f0031bae09f7cb508f866ec] [formerly 222e79ebc41a8f237f0031bae09f7cb508f866ec [formerly 6183d826ba62ec94ccfcb8f6e3b8d43e3e338703]]
Former-commit-id: e8dcc831cb13a1ac009fcf42c75ae66fe1f129fe
Former-commit-id: 13972ae517ad16b1ce806818aa2e0bf97bd70b20