config: warn on inaccessible files
Before reading a config file, we check "!access(path, R_OK)"
to make sure that the file exists and is readable. If it's
not, then we silently ignore it.

For the case of ENOENT, this is fine, as the presence of the
file is optional. For other cases, though, it may indicate a
configuration error (e.g., not having permissions to read
the file). Let's print a warning in these cases to let the
user know.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 24f347d971f9b494b2cffe87af2fb8d4fb35d9ec [formerly 24f347d971f9b494b2cffe87af2fb8d4fb35d9ec [formerly ba8bd8300a544959159f6bd3a7e03ac54f85ea3a]]
Former-commit-id: 48f9d4700cc5d591264eef7a3eedc70ff2fcafe5
Former-commit-id: cd4b8b89d212d844aaa9bad68926637261ba080a