gitignore: report access errors of exclude files
When we try to access gitignore files, we check for their
existence with a call to "access". We silently ignore
missing files. However, if a file is not readable, this may
be a configuration error; let's warn the user.

For $GIT_DIR/info/excludes or core.excludesfile, we can just
use access_or_warn. However, for per-directory files we
actually try to open them, so we must add a custom warning.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 609e1333ff64a230e5c3e7e8a594a7cab8fecfa4 [formerly 414654e9afa07d44351782567b4200242cd995a1] [formerly 414654e9afa07d44351782567b4200242cd995a1 [formerly 6966073102170171da4c25a4038bbd143fa6a75b]]
Former-commit-id: b29efcb947871f25a34c8510279b3ae0b95edaaa
Former-commit-id: ef550ef63250db7c1d234bd5cd82ba1a1a4e18be