warn_on_inaccessible(): a helper to warn on inaccessible paths
The previous series introduced warnings to multiple places, but it
could become tiring to see the warning on the same path over and
over again during a single run of Git.  Making just one function
responsible for issuing this warning, we could later choose to keep
track of which paths we issued a warning (it would involve a hash
table of paths after running them through real_path() or something)
in order to reduce noise.

Right now we do not know if the noise reduction is necessary, but it
still would be a good code reduction/sharing anyway.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f2bb735d373bf4132ce9a5a58b1f913d233fefb6 [formerly f2bb735d373bf4132ce9a5a58b1f913d233fefb6 [formerly 55b38a48e2a7ccfaaa7897a5fccb98327fa0e3c0]]
Former-commit-id: 1f1aa42bb751fab62ce441b215302203fa79bf74
Former-commit-id: 25ae5bda6216d483557f43c1de1bf56334551595