branch -v: align even when branch names are in UTF-8
Branch names are usually in ASCII so they are not the problem. The
problem most likely comes from "(no branch)" translation, which is
in UTF-8 and makes display-width calculation just wrong.  Clarify
this by renaming the field "len" in struct ref_item to "width", as
it stores the display-width and is used to compute the width of the
screen needed to show the names of all the branches, and compute the
display width using utf8_strwidth(), not byte-length with strlen().

Update document to mention the fact that we may want ref names in
UTF-8. Encodings that produce invalid UTF-8 are safe as utf8_strwidth()
falls back to strlen(). The ones that incidentally produce valid UTF-8
sequences will cause misalignment.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4490c072131fb9d68cd50d0271b2b589ea65c998 [formerly 4490c072131fb9d68cd50d0271b2b589ea65c998 [formerly 1452bd64f14efd888fe9c732f603173ba7fa595e]]
Former-commit-id: b564acee8ce2831ed986c37342f64a329c5ea23a
Former-commit-id: a348b9b536045691ae5136eec6f178ca2e7775ac