t5550: put auth-required repo in auth/dumb
In most of our tests, we put repos to be accessed by dumb
protocols in /dumb, and repos to be accessed by smart
protocols in /smart.  In our test apache setup, the whole
/auth hierarchy requires authentication. However, we don't
bother to split it by smart and dumb here because we are not
currently testing smart-http authentication at all.

That will change in future patches, so let's be explicit
that we are interested in testing dumb access here. This
also happens to match what t5540 does for the push tests.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4415cc99eb39efab7e64a40b0f7a4408f25d21b2 [formerly 4415cc99eb39efab7e64a40b0f7a4408f25d21b2 [formerly 726800a8b3bf3702490a79f0205e33b2a0dbbb64]]
Former-commit-id: 4bb8b20e57b3b7256e0a5a12826ff75f16fe28ac
Former-commit-id: 3c9050c4832c3a2f0241f462035db182322f019d