t/lib-httpd: only route auth/dumb to dumb repos
Our test apache config points all of auth/ directly to the
on-disk repositories via an Alias directive. This works fine
because everything authenticated is currently in auth/dumb,
which is a subset.  However, this would conflict with a
ScriptAlias for auth/smart (which will come in future
patches), so let's narrow the Alias.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 78686f56fea0fa9fa53917b473f70e78728ff0e8 [formerly b5550cd3ea592167b636174e6ff14accb55b9e6a] [formerly b5550cd3ea592167b636174e6ff14accb55b9e6a [formerly 05b577107dda131d46f93aa9bb7817c80bc30ee9]]
Former-commit-id: 63af60e0a2bf36c2860c5027dbd5235aec190031
Former-commit-id: 2d6ba0671d0ead9f421652552552d6d1f2288ffa