t/lib-httpd: recognize */smart/* repos as smart-http
We do not currently test authentication for smart-http repos
at all. Part of the infrastructure to do this is recognizing
that auth/smart is indeed a smart-http repo.

The current apache config recognizes only "^/smart/*" as
smart-http. Let's instead treat anything with /smart/ in the
URL as smart-http. This is obviously a stupid thing to do
for a real production site, but for our test suite we know
that our repositories will not have this magic string in the
name.

Note that we will route /foo/smart/bar.git directly to
git-http-backend/bar.git; in other words, everything before
the "/smart/" is irrelevant to finding the repo on disk (but
may impact apache config, for example by triggering auth
checks).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7dac9a93a49dd5821f9cecdc07962abb3ba06fb9 [formerly 7dac9a93a49dd5821f9cecdc07962abb3ba06fb9 [formerly 666aae9aed5a29019d2cd696d4258750c0dc96c7]]
Former-commit-id: d7c66952c2a5bf9f6f9e809433979fbb6811e6fc
Former-commit-id: 7a99469db1812865273f0dfb84ce5ec5d917e279