t: test http access to "half-auth" repositories
Some sites set up http access to repositories such that
fetching is anonymous and unauthenticated, but pushing is
authenticated. While there are multiple ways to do this, the
technique advertised in the git-http-backend manpage is to
block access to locations matching "/git-receive-pack$".

Let's emulate that advice in our test setup, which makes it
clear that this advice does not actually work.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b69a12e680e57536ad927882831edb7584e427fe [formerly b69a12e680e57536ad927882831edb7584e427fe [formerly 4c71009da60baee436358e84ff1057cd1c80e776]]
Former-commit-id: e1d687568a281823f6e4180992d60f2e895433d3
Former-commit-id: 68f6729b20f09602d8fbebbee05a4d7746c70925