checkout: reorder option handling
checkout operates in three different modes. On top of that it tries to
be smart by guessing the branch name for switching. This results in
messy option handling code. This patch reorders it so that

 - cmd_checkout() is responsible for parsing, preparing input and
   determining mode

 - Code of each mode is in checkout_paths() and checkout_branch(),
   where sanity checks are performed

Another slight improvement is always print branch name (or commit
name) when printing errors related ot them. This helps catch the case
where an option is mistaken as branch/commit.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 543cc138cbecd8dd104915fb37974fb68bdc895e [formerly 543cc138cbecd8dd104915fb37974fb68bdc895e [formerly b6312c27a06f31625d915d44531d5e728ffd8d63]]
Former-commit-id: 0f2650db6004d94ae08c129148c9cdf9a29b3c00
Former-commit-id: 29747fd570a34b46fb1e019b0dc8d83a54eda7b2