Make git-svn branch patterns match complete URL
When using the {word,[...]} style of configuration for tags and branches,
it appears the intent is to only match whole path parts, since the words
in the {} pattern are meta-character quoted.

When the pattern word appears in the beginning or middle of the url,
it's matched completely, since the left side, pattern, and (non-empty)
right side are joined together with path separators.

However, when the pattern word appears at the end of the URL, the
right side is an empty pattern, and the resulting regex matches
more than just the specified pattern.

For example, if you specify something along the lines of

    branches = branches/project/{release_1,release_2}

and your repository also contains "branches/project/release_1_2", you
will also get the release_1_2 branch.  By restricting the match regex
with anchors, this is avoided.

Signed-off-by: Ammon Riley <ammon.riley@gmail.com>
Signed-off-by: Eric Wong <normalperson@yhbt.net>

Former-commit-id: 736954827e814107db1dd0902061c415b4a569fb [formerly 5310ef45f844d33a812843ff45b1eefb9af8f2e8] [formerly 5310ef45f844d33a812843ff45b1eefb9af8f2e8 [formerly 9478b11968d61085bb4c4343395b6f52ce06380f]]
Former-commit-id: b2dfce20b414424f7dc36d01aa1987b284379f3b
Former-commit-id: e50ba148604fcbbc25c0690b4588e2dfa29d1738