fetch: use argv_array instead of hand-building arrays
Fetch invokes itself recursively when recursing into
submodules or handling "fetch --multiple". In both cases, it
builds the child's command line by pushing options onto a
statically-sized array. In both cases, the array is
currently just big enough to handle the largest possible
case. However, this technique is brittle and error-prone, so
let's replace it with a dynamic argv_array.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 65640b47619eea1ffa0e8fae7dc3b408a379e210 [formerly fedc19164ca3d80366ab49b0ba107e70ed2e32d9] [formerly fedc19164ca3d80366ab49b0ba107e70ed2e32d9 [formerly 85556d4e37db4c8ed88c68a602f6f85054996bea]]
Former-commit-id: 6d82dda65e6048fa98c8313984656809b9f5a4a7
Former-commit-id: 662192a2679cb4a8c1d844dd58d1859e7940d99b