argv-array: fix bogus cast when freeing array
Since the array struct stores a "const char **" argv member
(for compatibility with most of our argv-taking functions),
we have to cast away the const-ness when freeing its
elements.

However, we used the wrong type when doing so.  It doesn't
make a difference since free() take a void pointer anyway,
but it can be slightly confusing to a reader.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0066ba195bea1c10a564f8e4a5d06e468c81c473 [formerly 0066ba195bea1c10a564f8e4a5d06e468c81c473 [formerly ba4d1c7b1623b2c7ec198aee08036acf779375e6]]
Former-commit-id: 408406eaf36052744d5300a5ce9afd8185817bc6
Former-commit-id: 104529bdbfe2475a7ac0a2c3a199160c37c49321