submodule: use argv_array instead of hand-building arrays
fetch_populated_submodules() allocates the full argv array it uses to
recurse into the submodules from the number of given options plus the six
argv values it is going to add. It then initializes it with those values
which won't change during the iteration and copies the given options into
it. Inside the loop the two argv values different for each submodule get
replaced with those currently valid.

However, this technique is brittle and error-prone (as the comment to
explain the magic number 6 indicates), so let's replace it with an
argv_array. Instead of replacing the argv values, push them to the
argv_array just before the run_command() call (including the option
separating them) and pop them from the argv_array right after that.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: afbb2abd3ff58bd61a0837ed73144aedffe739a3 [formerly afbb2abd3ff58bd61a0837ed73144aedffe739a3 [formerly 50d89ad6542c8acafefa6d42f8b42dfa9b8fafe1]]
Former-commit-id: 2ee7022ea99d5bf4b548e6b41fb1552fc64750de
Former-commit-id: 6b6bca7c2de0ce9dbf9d676090892e4811fae245