test-lib.sh: Add check for invalid use of 'skip_all' facility
The 'skip_all' facility cannot be used after one or more tests
have been executed using (for example) 'test_expect_success'.
To do so results in invalid TAP output, which leads to 'prove'
complaining of "Parse errors: No plan found in TAP output".

Add a check for such invalid usage and abort the test with an
error message to alert the test author.

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b7824eb9f5044bfdac97418d19e2b6d432b04b90 [formerly b7824eb9f5044bfdac97418d19e2b6d432b04b90 [formerly bf4b7219322d007f3ebd2a002114956d025e4703]]
Former-commit-id: a6292154bbbbdaca866ee0ad313105b05e3a5ad2
Former-commit-id: 011e709a72912c8f87503e3fd6def6286eeaea77