path.c: Use vsnpath() in the implementation of git_path()
The current implementation of git_path() is essentially the same as
that of vsnpath(), with two minor differences. First, git_path()
currently insists that the git directory path is no longer than
PATH_MAX-100 characters in length. However, vsnpath() does not
attempt this arbitrary 100 character reservation for the remaining
path components. Second, vsnpath() uses the "is_dir_sep()" macro,
rather than comparing directly to '/', to determine if the git_dir
path component ends with a path separator.

In order to benefit from the above improvements, along with increased
compatability with git_snpath() and git_pathdup(), we reimplement the
git_path() function using vsnpath().

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 64eb44db85d251ada1e90b2577585fdb2f456ad2 [formerly 9c6bcc28823a4a4e913b989f1cdf5ad99a91e274] [formerly 9c6bcc28823a4a4e913b989f1cdf5ad99a91e274 [formerly 5c44252e13db2324d7c79d2b498886a925055111]]
Former-commit-id: 712b8d927ac97e53963afa309db2e3dce1913fa9
Former-commit-id: 33de0ce459bc84aa857143f2249109093623c6e5