remote: prefer subcommand name 'remove' to 'rm'
All remote subcommands are spelled out words except 'rm'. 'rm', being a
popular UNIX command name, may mislead users that there are also 'ls' or
'mv'. Use 'remove' to fit with the rest of subcommands.

'rm' is still supported and used in the test suite. It's just not
widely advertised.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3b570ac2af9f1826f96bbda157f657547a90fdf8 [formerly 7c72fa0fb25391d90bbf663955b71dc5aa94c2a5] [formerly 7c72fa0fb25391d90bbf663955b71dc5aa94c2a5 [formerly e17dba8fe15028425acd6a4ebebf1b8e9377d3c6]]
Former-commit-id: 07f50384d1f3c9c73d19f2363ff646056bcfbb2d
Former-commit-id: abe534a445fe9fa409ed31b9687672457bb15562