cvsimport: strip all inappropriate tag strings
Certain characters such as "?" can be present in a CVS tag name, but
git does not allow these characters in tags. If git-cvsimport
encounters a CVS tag that git cannot handle, cvsimport will error and
refuse to continue the import beyond that point.

When importing CVS tags, strip all the inappropriate strings from the
tag names as we translate them to git tag names.

Provide more debugging information to the user if we've altered the
tag and the "git tag" command still fails. Also, warn the user if we
end up skipping an (unusable) tag altogether.

Signed-off-by: Ken Dreyer <ktdreyer@ktdreyer.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e57ff99d706cd0e22d439d02b534fbc1a1b63000 [formerly e57ff99d706cd0e22d439d02b534fbc1a1b63000 [formerly 70b67b0792375c59f60f3e24f2d6757b24dc719c]]
Former-commit-id: 815ce853b49f44b6364fa810c86b1abd0226caa5
Former-commit-id: 8105f3556951ef69d26dfec0f9401246fb1c6610