t0060: verify that real_path() works correctly with absolute paths
There is currently a bug: if passed an absolute top-level path that
doesn't exist (e.g., "/foo") it incorrectly interprets the path as a
relative path (e.g., returns "$(pwd)/foo").  So mark the test as
failing.

These tests are skipped on Windows because test-path-utils operates on
a DOS-style absolute path even if a POSIX style absolute path is
passed as argument.

Adjusted for Windows by: Johannes Sixt <j6t@kdbg.org>

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c919d70a9569522636f58f5be418c83a8966a632 [formerly c919d70a9569522636f58f5be418c83a8966a632 [formerly 7bcf48dad83e7a07f27403c2ce9e5c29af9e317d]]
Former-commit-id: 8e7cf46cfd9d32e7aa242cc95dc6c516cfe442c5
Former-commit-id: eb846b5443302a7d77b777898415e25d4b2cdb26