fetch_pack(): update sought->nr to reflect number of unique entries
fetch_pack() removes duplicates from the "sought" list, thereby
shrinking the list.  But previously, the caller was not informed about
the shrinkage.  This would cause a spurious error message to be
emitted by cmd_fetch_pack() if "git fetch-pack" is called with
duplicate refnames.

Instead, remove duplicates using string_list_remove_duplicates(),
which adjusts sought->nr to reflect the new length of the list.

The last test of t5500 inexplicably *required* "git fetch-pack" to
fail when fetching a list of references that contains duplicates;
i.e., it insisted on the buggy behavior.  So change the test to expect
the correct behavior.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 980c8b0254fa02e1e2172844ce5ea47052384fe3 [formerly 1175e3c9b987bb8804d1ba01255663d117d5da3c] [formerly 1175e3c9b987bb8804d1ba01255663d117d5da3c [formerly 4c58f13ba67f9dbe67a351de61703a2db20712cc]]
Former-commit-id: 76bb2cd18c0e7ce47c468cd332d4c730af905555
Former-commit-id: c06f312a198b2554c0ea42b16e2e7a39215b1b74