filter_refs(): build refs list as we go
Instead of temporarily storing matched refs to temporary array
"return_refs", simply append them to newlist as we go.  This changes
the order of references in newlist to strictly sorted if "--all" and
"--depth" and named references are all specified, but that usage is
broken anyway (see the last two tests in t5500).

This changes the last test in t5500 from segfaulting into just
emitting a spurious error (this will be fixed in a moment).

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f4296f47cc41b5ded261068c20bea8e4c17ead3f [formerly 873a926f1b9f116d8499b4a3ce4b9745308986b1] [formerly 873a926f1b9f116d8499b4a3ce4b9745308986b1 [formerly 5096e4875303038ffc34af96c28cf2a2b7e103a6]]
Former-commit-id: 14837ed1fcc058044585fd71d4c7f6d7199d8e07
Former-commit-id: e0718a7bcdbea2f45356776b84a19c6751763064