fetch-pack: eliminate spurious error messages
It used to be that if "--all", "--depth", and also explicit references
were sought, then the explicit references were not handled correctly
in filter_refs() because the "--all --depth" code took precedence over
the explicit reference handling, and the explicit references were
never noted as having been found.  So check for explicitly sought
references before proceeding to the "--all --depth" logic.

This fixes two test cases in t5500.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4d94ef99434f2900b196fae0cf3e0c2962d13299 [formerly 4d94ef99434f2900b196fae0cf3e0c2962d13299 [formerly 5f0fc64513c8cb638e0d6010005d8e86c9dbca3a]]
Former-commit-id: 25b4bb4c8253333aed94a56eb0a81fbb5001b726
Former-commit-id: ece3b33699eee7c5c8aa6bde147967ce4cf947d4