t0060: split absolute path test in two to exercise some of it on Windows
Only the first half of the test works only on POSIX, the second half
passes on Windows as well.

A later test "real path removes other extra slashes" looks very similar,
but it does not make sense to split it in the same way: When two slashes
are prepended in front of an absolute DOS-style path on Windows, the
meaning of the path is changed (//server/share style), so that the test
cannot pass on Windows.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Acked-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 40cfc8a2d3eff8b19e67638f3539e1a5dbdc5330 [formerly 40cfc8a2d3eff8b19e67638f3539e1a5dbdc5330 [formerly bacca7852f2f304ebe3f579de1a6cf0152af3df1]]
Former-commit-id: 4060a0639a8e992a9fa35e5f8997fb6323bde961
Former-commit-id: cc2954d72502c570e7f10a8c20884d78904e36c8