git p4: move conflict prompt into run, add [q]uit input
When applying a commit to the p4 workspace fails, a prompt
asks what to do next.  This belongs up in run() instead
of in applyCommit(), where run() can notice, for instance,
that the prompt is unnecessary because this is the last commit.

Offer two options about how to continue at conflict: [s]kip or
[q]uit.  Having an explicit "quit" option gives git p4 a chance
to clean up, show the applied-commit summary, and do tag export.

Signed-off-by: Pete Wyckoff <pw@padd.com>
Acked-by: Luke Diamand <luke@diamand.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1eb76af8d5f16f29056b1f17a138b473151ee76d [formerly 1eb76af8d5f16f29056b1f17a138b473151ee76d [formerly 7e5dd9f2ccf2141d65e4621b765a136796b0d55b]]
Former-commit-id: 70187ee17c18fc87e69e430a5b5281452ce0242f
Former-commit-id: c0c0241d39befa541aafd08891fb345a4eabd4d4