git p4: revert deleted files after submit cancel
The user can decide not to continue with a submission,
by not saving the p4 submit template, then answering "no" to
the "Submit anyway?" prompt.  In this case, be sure to
return the p4 client to its initial state.

Deleted files were not reverted; fix this and test all cases.

Signed-off-by: Pete Wyckoff <pw@padd.com>
Acked-by: Luke Diamand <luke@diamand.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 673483ebdea7d0a99b676efe830d51d21c7b30f6 [formerly 673483ebdea7d0a99b676efe830d51d21c7b30f6 [formerly df9c5453b220923d7ec3e6291498d519596e23ed]]
Former-commit-id: f5b04edaad5ab39235b758019a7b78b2386f522e
Former-commit-id: 1da810a748457e1c9b3add4e9190aa9d5085596b