blame $path: avoid getting fooled by case insensitive filesystems
"git blame MAKEFILE" run in a history that has "Makefile" but not
MAKEFILE can get confused on a case insensitive filesystem, because
the check we run to see if there is a corresponding file in the
working tree with lstat("MAKEFILE") succeeds.  In addition to that
check, we have to make sure that the given path also exists in the
commit we start digging history from (i.e. "HEAD").

Note that this reveals the breakage in a test added in cd8ae20
(git-blame shouldn't crash if run in an unmerged tree, 2007-10-18),
which expects the entire merge-in-progress path to be blamed to the
working tree when it did not exist in our tree.  As it is clear in
the log message of that commit, the old breakage was that it was
causing an internal error and the fix was about avoiding it.

Just check that the command does not die an uncontrolled death.  For
this particular case, the blame should fail, as the history for the
file in that contents has not been committed yet at the point in the
test.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0efa37b21b221c4a136ea03a6b04b05c3b04f914 [formerly 0efa37b21b221c4a136ea03a6b04b05c3b04f914 [formerly ffcabccf5df17f12997feedafefeb5589b8c0511]]
Former-commit-id: 3af0d0f0ceadc573c54e4fe87e5ae5533627c101
Former-commit-id: 16254e6077f24627b389a7b15d03cd229e18a2e1