string_list: add two new functions for splitting strings
Add two new functions, string_list_split() and
string_list_split_in_place().  These split a string into a string_list
on a separator character.  The first makes copies of the substrings
(leaving the input string untouched) and the second splits the
original string in place, overwriting the separator characters with
NULs and referring to the original string's memory.

These functions are similar to the strbuf_split_*() functions except
that they work with the more powerful string_list interface.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 45b25e297f2433de554110331c638d19cd3ea3cf [formerly 45b25e297f2433de554110331c638d19cd3ea3cf [formerly ff919f965d20d003e3882c70de667f41a86349ac]]
Former-commit-id: e90c191500e6878a5792d85747ad4e78b96440c7
Former-commit-id: 3d92a6646951b15d32012d387db9a159dbe694e3