attr: failure to open a .gitattributes file is OK with ENOTDIR
Often we consult an in-tree .gitattributes file that exists per
directory.  Majority of directories do not usually have such a file,
and it is perfectly fine if we cannot open it because there is no
such file, but we do want to know when there is an I/O or permission
error.  Earlier, we made the codepath warn when we fail to open it
for reasons other than ENOENT for that reason.

We however sometimes have to attempt to open the .gitattributes file
from a directory that does not exist in the commit that is currently
checked out.  "git pack-objects" wants to know if a path is marked
with "-delta" attributes, and "git archive" wants to know about
export-ignore and export-subst attributes.  Both commands may and do
need to ask the attributes system about paths in an arbitrary
commit.  "git diff", after removing an entire directory, may want to
know textconv on paths that used to be in that directory.

Make sure we also ignore a failure to open per-directory attributes
file due to ENOTDIR.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4c8ec863f92d2d3a3b975cc9679e3b0cd8a17c26 [formerly 4c8ec863f92d2d3a3b975cc9679e3b0cd8a17c26 [formerly 8e950dab86eafa8f4ea671fab593ca6210bc2e15]]
Former-commit-id: bbcfc99264d80b3bb7678fb0436cc80110f1a929
Former-commit-id: f6bd89307f51bcb87ac5f806544b8dccab4b33a6