grep: teach --debug option to dump the parse tree
Our "grep" allows complex boolean expressions to be formed to match
each individual line with operators like --and, '(', ')' and --not.
Introduce the "--debug" option to show the parse tree to help people
who want to debug and enhance it.

Also "log" learns "--grep-debug" option to do the same.  The command
line parser to the log family is a lot more limited than the general
"git grep" parser, but it has special handling for header matching
(e.g. "--author"), and a parse tree is valuable when working on it.

Note that "--all-match" is *not* any individual node in the parse
tree.  It is an instruction to the evaluator to check all the nodes
in the top-level backbone have matched and reject a document as
non-matching otherwise.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 87366f6f5d748b448f2908f6c541a2e07b3bab1b [formerly 87366f6f5d748b448f2908f6c541a2e07b3bab1b [formerly 17bf35a3c7b46df7131681bcc5bee5f12e1caec4]]
Former-commit-id: cbe91a907bb24570e8c968b8164755ce990c5a0a
Former-commit-id: 2759f4360dc998e26fc1a15736c6450f6c836c93