Documentation: describe subject more precisely
The discussion of email subject throughout the documentation is
misleading; it indicates that the first line will always become
the subject.  In fact, the subject is generally all lines up until
the first full blank line.

This patch refines that, and makes more use of the concept of a
commit title, with the title being all text up to the first blank line.

Signed-off-by: Jeremy White <jwhite@codeweavers.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ff06240cb0f46bc5f8f9f53fe372d39b76adce2a [formerly ff06240cb0f46bc5f8f9f53fe372d39b76adce2a [formerly 52ffe995b9a3fc43de8eca80bedab2e055aac562]]
Former-commit-id: 41421bce3ebb82d78b8d5e14205b8d18eda8ea09
Former-commit-id: d2ed3cc62aefe86af672284b2088b1637068bb74