MALLOC_CHECK: various clean-ups
The most important in this change is to avoid affecting anything
when test-lib is used from perf-lib.  It also limits the effect of
the MALLOC_CHECK only to what is run inside the actual test, and
uses a fixed MALLOC_PERTURB_ in order to avoid hurting repeatability
of the tests.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8cae7bb05fcd1ba8fa4a1f93fd65ca902693a109 [formerly 6cfd60aefbb542df46c5e4f969ac8c6f87141b06] [formerly 6cfd60aefbb542df46c5e4f969ac8c6f87141b06 [formerly 1b3185fc2be3cc92a834fcfcb7110d1072184700]]
Former-commit-id: 4b6d123341d72244b63157cfa7a92a7a07ab36fd
Former-commit-id: e0d720434ecada9b77359cbfe137a9e6f46c88f8