string_list API: document what "sorted" means
The recent work on using string_list to represent the list of refs
that matched with the refs on the other side during fetch heavily
depends on the sort order by string_list's implementation, and
changing string_list will break it.  Document that it uses strcmp()
order, at least for now.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 10e2be4087cee63807928400b23c4751747aa9e9 [formerly 10e2be4087cee63807928400b23c4751747aa9e9 [formerly 1959bf6430fda685c46241f220f521a01f9469b7]]
Former-commit-id: 6e3d531d60dfa6dbd265094b633429eab9ae775b
Former-commit-id: 63557a560389b48bbc2f17356538e0b05386ccc8