mailinfo: do not concatenate charset= attribute values from mime headers
"Content-type: text/plain; charset=UTF-8" header should not appear
twice in the input, but it is always better to gracefully deal with
such a case.  The current code concatenates the value to the values
we have seen previously, producing nonsense such as "utf8UTF-8".

Instead of concatenating, forget the previous value and use the last
value we see.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cdeae2233ae9ba4bfac6f15513f96c93c04c5755 [formerly cdeae2233ae9ba4bfac6f15513f96c93c04c5755 [formerly 176943b9658b1fa9ae3972156100d831d19c3576]]
Former-commit-id: ca3739a5421fb2bf46fa4d698e5ae5ad2089e9da
Former-commit-id: 410c7d41f690190c9b21dfca2a631cf3c9691505