poll() exits too early with EFAULT if 1st arg is NULL
If poll() is used as a milli-second sleep, like in help.c, by passing a NULL
in the 1st and a 0 in the 2nd arg, it exits with EFAULT.

As per Paolo Bonzini, the original author, this is a bug and to be fixed
Like in this commit, which is not to exit if the 2nd arg is 0. It got fixed
In gnulib in the same manner the other day.

Signed-off-by: Joachim Schmitz <jojo@schmitz-digital.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2ab552e4521283c645c8dde2e12674388ba709fb [formerly 2ab552e4521283c645c8dde2e12674388ba709fb [formerly 32fde6575e4139710ea5259ee9b7c750b7c2dde0]]
Former-commit-id: 8fc0fd3f8d25db70d49ae34dd9aa6fe0bd169ab4
Former-commit-id: 4070b4698a7fca1befb8bf6043a41b6c7e3f1ed4