Documentation/git-filter-branch: Move note about effect of removing commits
The note that explains that changes introduced by removed commits are
preserved should be placed directly after the paragraph that describes
such commits removal.  Otherwise the reference to "the commits" appears
out of context.

Also the big example that follows "Consider this history" is about
rewriting part of the history DAG.  Move the paragraph that
describes the operation close to it.

Signed-off-by: Andreas Schwab <schwab@linux-m68k.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d368a5a15df0c726d924a08c7e24081a7a738bc3 [formerly d368a5a15df0c726d924a08c7e24081a7a738bc3 [formerly 8093ae8854a50edf44e47c0390d3021ffd07bc00]]
Former-commit-id: 7e7556918d9f02070ba4b1b0123eb5bd807127c5
Former-commit-id: 4661a3aff5f0061100b119f13f0426188ec054db