remote-svn, vcs-svn: Enable fetching to private refs
The reference to update by the fast-import stream is hard-coded.  When
fetching from a remote the remote-helper shall update refs in a
private namespace, i.e. a private subdir of refs/.  This namespace is
defined by the 'refspec' capability, that the remote-helper advertises
as a reply to the 'capabilities' command.

Extend svndump and fast-export to allow passing the target ref.
Update svn-fe to be compatible.

Signed-off-by: Florian Achleitner <florian.achleitner.2.6.31@gmail.com>
Acked-by: David Michael Barr <b@rr-dav.id.au>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ecdfbbb9580f3b931b32359a98bee5b61de70dab [formerly ecdfbbb9580f3b931b32359a98bee5b61de70dab [formerly 271fd1fc2a25ebfc1782aab0d39b613555d22315]]
Former-commit-id: d08085bc9fa78a2b9da13808e7b64145e7ad2f8f
Former-commit-id: 5c0d255263ef2fd9352e5da73e6ad5d57facf4ab