Allow reading svn dumps from files via file:// urls
For testing as well as for importing large, already available dumps,
it's useful to bypass svnrdump and replay the svndump from a file
directly.

Add support for file:// urls in the remote url, e.g.

  svn::file:///path/to/dump

When the remote helper finds an url starting with file:// it tries to
open that file instead of invoking svnrdump.

Signed-off-by: Florian Achleitner <florian.achleitner.2.6.31@gmail.com>
Acked-by: David Michael Barr <b@rr-dav.id.au>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 96067ebaba2a7a8d1fb65a6f0d6c4cd48983acaf [formerly 96067ebaba2a7a8d1fb65a6f0d6c4cd48983acaf [formerly f6529de9f4fc66de7679fa1204da1da50f6292fb]]
Former-commit-id: d575f6fd8ce8cfd0f800ff39402dd0898855146d
Former-commit-id: 0f20106d0809220b634254dded37ec86a9e266e6