Add a svnrdump-simulator replaying a dump file for testing
To ease testing without depending on a reachable svn server, this
compact python script mimics parts of svnrdumps behaviour.  It
requires the remote url to start with sim://.

Start and end revisions are evaluated.  If the requested revision
doesn't exist, as it is the case with incremental imports, if no new
commit was added, it returns 1 (like svnrdump).

To allow using the same dump file for simulating multiple incremental
imports, the highest revision can be limited by setting the environment
variable SVNRMAX to that value. This simulates the situation where
higher revs don't exist yet.

Signed-off-by: Florian Achleitner <florian.achleitner.2.6.31@gmail.com>
Acked-by: David Michael Barr <b@rr-dav.id.au>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d98e071bf7d154b8d021bf90468c3dfcec93b632 [formerly d98e071bf7d154b8d021bf90468c3dfcec93b632 [formerly 16a7185447e28312d5e316b92b7d3d37cb25b079]]
Former-commit-id: 1ebb56032b50dfc441eb957afc9d6a7b63ea10fe
Former-commit-id: c525eb94566365dafe6954dc5f23b64ee7876bad