remote-svn: add marks-file regeneration
fast-import mark files are stored outside the object database and are
therefore not fetched and can be lost somehow else.  marks provide a
svn revision --> git sha1 mapping, while the notes that are attached
to each commit when it is imported provide a git sha1 --> svn revision
mapping.

If the marks file is not available or not plausible, regenerate it by
walking through the notes tree.  , i.e.  The plausibility check tests
if the highest revision in the marks file matches the revision of the
top ref. It doesn't ensure that the mark file is completely correct.
This could only be done with an effort equal to unconditional
regeneration.

Signed-off-by: Florian Achleitner <florian.achleitner.2.6.31@gmail.com>
Acked-by: David Michael Barr <b@rr-dav.id.au>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2000d5c0633f3d13ce179fdedc38f4a32f42cc5f [formerly 2000d5c0633f3d13ce179fdedc38f4a32f42cc5f [formerly 5bfc76b5b275bbf0ddc2cf1a3a33fa6156addbd8]]
Former-commit-id: a38a8b2e93ccfad7ec2a930ae2b10153b8fa1dac
Former-commit-id: 084b4ce5289c4290eef22da0d6d7d6c670f521d8