completion: fix shell expansion of items
As reported by Jeroen Meijer[1]; the current code doesn't deal properly
with items (tags, branches, etc.) that have ${} in them because they get
expaned by bash while using compgen.

A simple solution is to quote the items so they get expanded properly
(\$\{\}).

In order to achieve that I took bash-completion's quote() function,
which is rather simple, and renamed it to __git_quote() as per Jeff
King's suggestion.

Solves the original problem for me.

[1] http://article.gmane.org/gmane.comp.version-control.git/201596

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Acked-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fb42fa450fdab8f2d70d19e105dc07a742216282 [formerly fb42fa450fdab8f2d70d19e105dc07a742216282 [formerly 25ae7cfd19c8f21721363c64163cd5d9d1135b20]]
Former-commit-id: e96772429f78edca8b39af1aaeb476a951decdaa
Former-commit-id: c8e71fc2436d087c29ce8a518d5069b2a5c1d1c6