clone --single: limit the fetch refspec to fetched branch
After running "git clone --single", the resulting repository has the
usual default "+refs/heads/*:refs/remotes/origin/*" wildcard fetch
refspec installed, which means that a subsequent "git fetch" will
end up grabbing all the other branches.

Update the fetch refspec to cover only the singly cloned ref instead
to correct this.

That means:
If "--single" is used without "--branch" or "--mirror", the
fetch refspec covers the branch on which remote's HEAD points to.
If "--single" is used with "--branch", it'll cover only the branch
specified in the "--branch" option.
If "--single" is combined with "--mirror", then it'll cover all
refs of the cloned repository.
If "--single" is used with "--branch" that specifies a tag, then
it'll cover only the ref for this tag.

Signed-off-by: Ralf Thielow <ralf.thielow@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e10f428eaec17d8b4fed7517524d3f447a84148a [formerly 916a766bf3a4d1597f7414d4fd5452e47cdab581] [formerly 916a766bf3a4d1597f7414d4fd5452e47cdab581 [formerly 31b808a03290a4f87c571cc1c61f26d8a03c7025]]
Former-commit-id: eb5e9066df16d7f687fb2bd0c172aa51f1056c68
Former-commit-id: bea953bbedc20c9c3be573aa068d1e0042c32e32