git blame: document that it always follows origin across whole-file renames
Make it clear to people who (rightly or wrongly) think that the
"--follow" option should follow origin across while-file renames
that we already do so.  That would explain the output that they see
when they do give the "--follow" option to the command.

We may or may not want to do a "--no-follow" patch as a follow-up,
but that is a separate topic.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f5bf5bf57fca79286d64cc066a55930362e4e965 [formerly f5bf5bf57fca79286d64cc066a55930362e4e965 [formerly e5dce96e9e26f8e30291c79fa5647313c64b2150]]
Former-commit-id: 83e69fb82d7ba78c0c91bb6592e147b8d148ea22
Former-commit-id: ea86ed8164435a8c1fd0a9a31fabbcf571535061