blame: pay attention to --no-follow
If you know your history did not have renames, or if you care only
about the history after a large rename that happened some time ago,
"git blame --no-follow $path" is a way to tell the command not to
bother about renames.

When you use -C, the lines that came from the renamed file will
still be found without the whole-file rename detection, so it is not
all that interesting either way, though.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1b1bc927a9314b446cce8b554a0a1fb467615713 [formerly 1b1bc927a9314b446cce8b554a0a1fb467615713 [formerly 3d1aa56671e7f33740dcd2cf0642d5ce09bd6a95]]
Former-commit-id: 5d11adabc35be05174fa906c19a737cd0cc72fc8
Former-commit-id: 391c297b4e89c1eee7aeea03addc220f327f73fc