test-lib: use $SHELL_PATH, not $SHELL
The codepath for handling "--tee" ends up relaunching the test
script under a shell, and that one has to be a Bourne.  But we
incorrectly used $SHELL, which could be a non-Bourne (e.g. zsh or
csh); we have the Makefile variable $SHELL_PATH for exactly that,
so use it instead.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6a16fed9553feac6193c49f5f5ff0efff83f0e8e [formerly aed2d2c65971e2d21d2d44b8f29961acf66597f9] [formerly aed2d2c65971e2d21d2d44b8f29961acf66597f9 [formerly 4cde519fe666af478be4b86659c518949a853f91]]
Former-commit-id: c9b7ad937c5fdbc4da868e695643e29ac055d22d
Former-commit-id: 26f7df6c316b8ed163c8cadff1d1fccba1ad1dbf