gitk: Work around empty back and forward images when buttons are disabled
On Mac, the back and forward buttons show an empty rectange instead of
a grayed-out arrow when they are disabled. The reason is a Tk bug on Mac
that causes disabled images not to draw correctly (not to draw at all,
that is); see
<https://groups.google.com/forum/?fromgroups=#!topic/comp.lang.tcl/V-nW1JBq0eU>.

To work around this, we explicitly provide gray images for the disabled
state; I think this looks better than the default stipple effect that you
get on Windows as well, but that may be a matter of taste.

Signed-off-by: Stefan Haller <stefan@haller-berlin.de>
Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: 2bfeed8fc79aa3f9fa7a329042ea7bee2b153a03 [formerly c57c12dbc8c6db4f67af80f6cdde79e48a4c4aaa] [formerly c57c12dbc8c6db4f67af80f6cdde79e48a4c4aaa [formerly f062e50fe69e01376b6ebf7ade8402b782620ac3]]
Former-commit-id: d905753aa342a171eade9e529b32cbb035381555
Former-commit-id: 4a4f4cb8cd5cd6e72a9102cab6b520f9a043e8a3