submodule: if $command was not matched, don't parse other args
"git submodule" command DWIMs the command line and assumes a
unspecified action word for 'status' action.  This is a UI mistake
that leads to a confusing behaviour.  A mistyped command name is
instead treated as a request for 'status' of the submodule with that
name, e.g.

    $ git submodule show
    error: pathspec 'show' did not match any file(s) known to git.
    Did you forget to 'git add'?

Stop DWIMming an unknown or mistyped subcommand name as pathspec
given to unspelled "status" subcommand.  "git submodule" without any
argument is still interpreted as "git submodule status", but its
value is questionable.

Adjust t7400 to match, and stop advertising the default subcommand
being 'status' which does not help much in practice, other than
promoting laziness and confusion.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 269f212cf31dee9b0f9bda168dee7ef7358a182d [formerly 0b9727fe94e866539a3bf4a4afb8927e78046676] [formerly 0b9727fe94e866539a3bf4a4afb8927e78046676 [formerly af9c9f97137dfb3f20a17d76520803c0312fd3e4]]
Former-commit-id: 109045aefdb420ce45f735bd5967f0cbc4820762
Former-commit-id: c49ec3b71a26bf4bf48f20ea4fc3ae7bbe6cebe1