archive-zip: write extended timestamp
File modification times in ZIP files are encoded in DOS format: local
time with a granularity of two seconds.  Add an extra field to all
archive entries to also record the mtime in Unix' fashion, as UTC with
a granularity of one second.

This has the desirable side-effect of convincing Info-ZIP unzip 6.00
to respect general purpose flag 11, which is used to indicate that a
file name is encoded in UTF-8.  Any extra field would do, actually,
but the extended timestamp is a reasonably small one (22 bytes per
entry).  Archives created by Info-ZIP zip 3.0 contain it, too (but
with ctime and atime as well).

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6e5ace2626c38552964b0dbe1eee679a8b36d879 [formerly 6e5ace2626c38552964b0dbe1eee679a8b36d879 [formerly 227bf59806dc408d60f8f80a14d8dfe71b92c6d2]]
Former-commit-id: 51b431293febeb3c9736cf572d1c693f29d51f04
Former-commit-id: e62e293b73ff10e35587df45e69519d394825934