mergetool--lib: Allow custom commands to override built-ins
Allow users to override the default commands provided by the
mergetools/* scriptlets.

Users occasionally run into problems where they expect to be
able to override the built-in tool names.  The documentation
does not explicitly mention that built-ins cannot be overridden,
so it's easy to assume that it should work.

Lift this restriction so that built-in tools are handled the
same way as user-configured tools.  Add tests to guarantee this
behavior.

A nice benefit of this change is that it protects users from
having future versions of git trump their custom configuration
with a new built-in tool.

C.f.:

http://stackoverflow.com/questions/7435002/mergetool-from-gitconfig-being-ignored
http://thread.gmane.org/gmane.comp.version-control.msysgit/13188
http://thread.gmane.org/gmane.comp.version-control.git/148267

Signed-off-by: David Aguilar <davvid@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 06709a29e93379caeff49c48eb4536af6708a455 [formerly 06709a29e93379caeff49c48eb4536af6708a455 [formerly a427ef7acc9d932d1c203dd2fae67f51c4b1d1e8]]
Former-commit-id: 6f114e8f59a72bb52c1944fac2a5dff7ecb222b8
Former-commit-id: c534169ffdb6bc0e0af0875e31023ee263ae8fad