grep: prepare for new header field filter
grep supports only author and committer headers, which have the same
special treatment that later headers may or may not have. Check for
field type and only strip_timestamp() when the field is either author
or committer.

GREP_HEADER_FIELD_MAX is put in the grep_header_field enum to be
calculated automatically, correctly, as long as it's at the end of the
enum.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 023caa6e095c04536a1c581ac332fe6e24e675e1 [formerly 023caa6e095c04536a1c581ac332fe6e24e675e1 [formerly ad4813b3c2513c5dc7e84305ab8a393b32124977]]
Former-commit-id: e5c56e01d4f8b42dc9f071db8d80d909aad62886
Former-commit-id: 40cac76d1f932fc899588f265940fb57b4cd07c2