t4055: avoid use of sed 'a' command
The 'a', 'i' and 'c' commands take a literal text to be added
followed by backslash, but then in the source we cannot indent
the literal text which makes it ugly.

We need to also remember to double the backslash inside double
quotes.

Avoid these issues altogether by having an extra line in a template
file and generate test vectors by deleting the line or replacing the
line and not using the 'a' command.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4508b6affcb3dfd2427ef37e320a0b119c407c7a [formerly 4508b6affcb3dfd2427ef37e320a0b119c407c7a [formerly 50fb51e7e81748182a770d3943ec5730357b27e7]]
Former-commit-id: 480253bc96191b669aea71a82f858d79c44b3f5f
Former-commit-id: 743dc0265bd91d8617b13a49c82d4101d42a53b2