log --grep: accept --basic-regexp and --perl-regexp
When we added the "--perl-regexp" option (or "-P") to "git grep", we
should have done the same for the commands in the "git log" family,
but somehow we forgot to do so.  This corrects it, but we will
reserve the short-and-sweet "-P" option for something else for now.

Also introduce the "--basic-regexp" option for completeness, so that
the "last one wins" principle can be used to defeat an earlier -E
option, e.g. "git log -E --basic-regexp --grep='<bre>'".  Note that
it cannot have the short "-G" option as the option is to grep in the
patch text in the context of "log" family.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9aef476548d7a95d1fc97905b273577a77eb902e [formerly 9aef476548d7a95d1fc97905b273577a77eb902e [formerly 727b6fc3edacebcde0a3dc849af95bd7f5a4d4a1]]
Former-commit-id: 9fa434056f627eead7e96ddaf1a89ab32e8ab707
Former-commit-id: 20a1fa3509e576b0f1525397c4cb957bae389a08