MALLOC_CHECK: Allow checking to be disabled from config.mak
The malloc checks can be disabled using the TEST_NO_MALLOC_CHECK
variable, either from the environment or command line of an
'make test' invocation. In order to allow the malloc checks to be
disabled from the 'config.mak' file, we add TEST_NO_MALLOC_CHECK
to the environment using an export directive.

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0f3f212874539f8ab4dde217b6d75b92cd091a0d [formerly 0f3f212874539f8ab4dde217b6d75b92cd091a0d [formerly 1266686b97d9c1783b89911a938bf992d8292bc6]]
Former-commit-id: b7d7c90cd8da1fa55b94b8b01089c3e6b07189e3
Former-commit-id: 6aacafc859d50f807c904553ed311157a6f066a8