gitk: Fix crash when not using themed widgets
When configured not to use themed widgets gitk may crash on launch with
a message that says that the image "bm-left disabled bm-left-gray"
doesn't exist. This happens when the left and right arrow buttons are
created.

The crash can be avoided by configuring the buttons differently
depending on whether or not themed widgets are used. If themed widgets
are not used then only set the images to bm-left and bm-right
respectively, and keep the old behavior when themed widgets are used.

The previous behaviour was added in f062e50f to work around a bug in Tk
on OS X where the disabled state did not display properly. The buttons
may still not display correctly, however the workaround added in
f062e50f will still apply if gitk is used with themed widgets.

Make gitk not crash on launch when not using themed widgets.

Signed-off-by: Marcus Karlsson <mk@acc.umu.se>
Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: f74a711be8a4768af5b13f85d5a9b09f54f311dc [formerly f74a711be8a4768af5b13f85d5a9b09f54f311dc [formerly 62e9ac5edfd7d9f24482ce767eec4b1ac96db4a4]]
Former-commit-id: 742e64a922c46112afe62a4a7923ddd11cf7123e
Former-commit-id: b40d21429682a40ee2cdb6b1ad46b7751e5701f8