Allow __git_ps1 to be used in PROMPT_COMMAND
Changes __git_ps1 to allow its use as PROMPT_COMMAND in bash
in addition to setting PS1 with __git_ps1 in a command substitution.
PROMPT_COMMAND has advantages for using color without running
into prompt-wrapping issues. Only by assigning \[ and \] to PS1
directly can bash know that these and the enclosed zero-width codes in
between don't count in the length of the prompt.

Signed-off-by: Simon Oosthoek <s.oosthoek@xs4all.nl>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d5c34aadc5b2a7c18098ff0aacb6b351cca88294 [formerly d5c34aadc5b2a7c18098ff0aacb6b351cca88294 [formerly 1bfc51ac814125de03ddf1900245e42d6ce0d250]]
Former-commit-id: 8190c2649dee803844cecc514737a420deb6ce3c
Former-commit-id: b3d2ca91d41f05d9d81daa24afb6cac1cc21f07f