rebase: Handle cases where format-patch fails
'format-patch' could fail due to reasons such as out of memory. Such
failures are not detected or handled, which causes rebase to incorrectly
think that it completed successfully and continue with cleanup. i.e.
calling move_to_original_branch

Instead of using a pipe, we separate 'format-patch' and 'am' by using an
intermediate file. This gurantees that we can invoke 'am' with the
complete input, or not invoking 'am' at all if 'format-patch' failed.

Also remove the use of '&&' at the end of the if-block, and rearrange
the 'write_basic_state' and 'move_to_original_branch' to make the logic
flow a bit better and easier to read.

Signed-off-by: Andrew Wong <andrew.kw.w@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 87ba761fb16bbfa0f801c4d936594801dfac089b [formerly 87ba761fb16bbfa0f801c4d936594801dfac089b [formerly e481af06be6fa112cec9f53fabd51d3b4cc6c64d]]
Former-commit-id: 9ffe638d72c222c46090e9f10bc1e2152cdc4332
Former-commit-id: 886e24c8826710ef149d4e959cc13642f103c31a