remote-curl: do not call run_slot repeatedly
Commit b81401c (http: prompt for credentials on failed POST)
taught post_rpc to call run_slot in a loop in order to retry
a request after asking the user for credentials. However,
after a call to run_slot we will have called
finish_active_slot. This means we have released the slot,
and we should no longer look at it.

As it happens, this does not cause any bugs in the current
code, since we know that we are not using curl_multi in this
code path, and therefore nobody will have taken over our
slot in the meantime. However, it is good form to actually
call get_active_slot again. It also future proofs us against
changes in the http code.

We can do this by jumping back to a retry label at the top
of our function. We just need to reorder a few setup lines
that should not be repeated; everything else within the loop
is either idempotent, needs to be repeated, or in a path we
do not follow (e.g., we do not even try when large_request
is set, because we don't know how much data we might have
streamed from our helper program).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ff34bbe4a11062a3254518a9340b5d7f3a1bf8b1 [formerly ff34bbe4a11062a3254518a9340b5d7f3a1bf8b1 [formerly abf8df869c79ee6fa021b117e60683fe149131d8]]
Former-commit-id: f2d28ee3a9fc57a319a29eb2197a2e06a8747ac1
Former-commit-id: f83d2a26c437dc66c03ddd95ade3d328fd51c20a