grep: stop looking at random places for .gitattributes
grep searches for .gitattributes using "name" field in struct
grep_source but that field is not real on-disk path name. For example,
"grep pattern rev" fills the field with "rev:path", and Git looks for
.gitattributes in the (non-existent but exploitable) path "rev:path"
instead of "path".

This patch passes real paths down to grep_source_load_driver() when:

 - grep on work tree
 - grep on the index
 - grep a commit (or a tag if it points to a commit)

so that these cases look up .gitattributes at proper paths.
.gitattributes lookup is disabled in all other cases.

Initial-work-by: Jeff King <peff@peff.net>
Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2564c5e18dae2fdf25c9a750d6e0aedf6af42b83 [formerly 2564c5e18dae2fdf25c9a750d6e0aedf6af42b83 [formerly 55c61688ea1e41f4a8c26f957bf1bc43cd39ed97]]
Former-commit-id: 34c3fa83005b8acec3cf7e25af01e5ab2c4d5565
Former-commit-id: f62f8e82da7fc2db1b7c38eaaad29571b673a818