config: treat user and xdg config permission problems as errors
Git reads multiple configuration files: settings come first from the
system config file (typically /etc/gitconfig), then the xdg config
file (typically ~/.config/git/config), then the user's dotfile
(~/.gitconfig), then the repository configuration (.git/config).

Git has always used access(2) to decide whether to use each file; as
an unfortunate side effect, that means that if one of these files is
unreadable (e.g., EPERM or EIO), git skips it.  So if I use
~/.gitconfig to override some settings but make a mistake and give it
the wrong permissions then I am subject to the settings the sysadmin
chose for /etc/gitconfig.

Better to error out and ask the user to correct the problem.

This only affects the user and xdg config files, since the user
presumably has enough access to fix their permissions.  If the system
config file is unreadable, the best we can do is to warn about it so
the user knows to notify someone and get on with work in the meantime.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2bd9eb61ac116985a6b1ca923986c6b5761c4f0e [formerly 49725ad2ad3462429722105d94cf468b070cf7ed] [formerly 49725ad2ad3462429722105d94cf468b070cf7ed [formerly 96b9e0e313604f77456906ce58db8f366e47f2ab]]
Former-commit-id: 1ac0b46dc320feb421824c1c6bde4fbbae31e1ee
Former-commit-id: 4ac61de6eb9c6e2401ab738e1f5427f6f2d3a48c