exclude: fix a bug in prefix compare optimization
When "namelen" becomes zero at this stage, we have matched the fixed
part, but whether it actually matches the pattern still depends on the
pattern in "exclude". As demonstrated in t3001, path "three/a.3"
exists and it matches the "three/a.3" part in pattern "three/a.3[abc]",
but that does not mean a true match.

Don't be too optimistic and let fnmatch() do the job.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 533c439b795b5f52bd2154a4c9ffa92935cce11a [formerly 4260fae706611f0d54de728fb23e1977f97bb0d2] [formerly 4260fae706611f0d54de728fb23e1977f97bb0d2 [formerly a3ea4d7199870595eb3a264c4b41e725b4417bc2]]
Former-commit-id: 32a1683dc237c175b38ab5644ca9eec2ebca265b
Former-commit-id: 6625f25cda0ff7c3cc140398bc2b04cd5b1b36f6