refs: lock symref that is to be deleted, not its target
When delete_ref is called on a symref then it locks its target and then
either deletes the target or the symref, depending on whether the flag
REF_NODEREF was set in the parameter delopt.

Instead, simply pass the flag to lock_ref_sha1_basic, which will then
either lock the target or the symref, and delete the locked ref.

This reimplements part of eca35a25 (Fix git branch -m for symrefs.).

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: af2a4c47a2ce65aac81cca80ef844e97d9979dd3 [formerly 440446238cdcab3623f26ab44555f7ddf29879e0] [formerly 440446238cdcab3623f26ab44555f7ddf29879e0 [formerly 547d058fb6fefad147675a6d3fdb9439c57916a1]]
Former-commit-id: a374f0599a34bb4c60710691a1c7fd0fca0f92cf
Former-commit-id: b8fcac24b226e5b5d6b689f93ec26fc1cc9d93d7