maybe_flush_or_die: move a too-loose Windows specific error
 check to compat

Commit b2f5e268 (Windows: Work around an oddity when a pipe with no reader
is written to) introduced a check for EINVAL after fflush() to fight
spurious "Invalid argument" errors on Windows when a pipe was broken. But
this check may hide real errors on systems that do not have the this odd
behavior. Introduce an fflush wrapper in compat/mingw.* so that the treatment
is only applied on Windows.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4b53875debd9c2da806349700f185d08921f7506 [formerly 59f41c4799eb566c9e3da8d8d333bf42c4472d58] [formerly 59f41c4799eb566c9e3da8d8d333bf42c4472d58 [formerly 84adb641545f4b58f9276adf099f840ea2928e44]]
Former-commit-id: 40f4e309f34ac2ffb0cebf08199e80f4c40a2117
Former-commit-id: 08a009d18c2a67b1d7b07523d794d2986c8707db