pretty_print_commit(): do not append notes message
The only case pretty_print_commit() appends notes message to the log
message taken from the commit is when show_log() calls it with the
notes_message field set, and the output format is not the userformat
(i.e. when substituting "%N").  No other users of this function sets
this field in the pretty_print_context, as can be easily verified in
the previous step.

Hoist the code to append the notes message to the caller.

Up to this point, no functionality change is intended.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 21baae0c6ae7ed9305d5d77a26461b57fc07d3b3 [formerly 21baae0c6ae7ed9305d5d77a26461b57fc07d3b3 [formerly 5a664cf2c76ed83166f5f3ce7385e3cb058e8834]]
Former-commit-id: 5a758d9e2ba5306ad91eba97f5eb5692f6163b4e
Former-commit-id: 80aa2214e2f5cc7b3c3c07aef14da6e3ef463616