git-sh-setup: refactor ident-parsing functions
The only ident-parsing function we currently provide is
get_author_ident_from_commit. This is not very
flexible for two reasons:

  1. It takes a commit as an argument, and can't read from
     commit headers saved on disk.

  2. It will only parse authors, not committers.

This patch provides a more flexible interface which will
parse multiple idents from a commit provide on stdin. We can
easily use it as a building block for the current function
to retain compatibility.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f61027a7d709b2b0a31cf49421ff716ec081ec1e [formerly f61027a7d709b2b0a31cf49421ff716ec081ec1e [formerly ce80ca566af713c85ce3a1c20b466486058a971a]]
Former-commit-id: 591ea918b8093c25304b82ae5f0fe9a036d25513
Former-commit-id: 666226e8868d72f09cba30b151e6598680bf9300