filter-branch: use git-sh-setup's ident parsing functions
This saves us some code, but it also reduces the number of
processes we start for each filtered commit. Since we can
parse both author and committer in the same sed invocation,
we save one process. And since the new interface avoids tr,
we save 4 processes.

It also avoids using "tr", which has had some odd
portability problems reported with from Solaris's xpg6
version.

We also tweak one of the tests in t7003 to double-check that
we are properly exporting the variables (because test-lib.sh
exports GIT_AUTHOR_NAME, it will be automatically exported
in subprograms. We override this to make sure that
filter-branch handles it properly itself).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 93e92eb7974e1359abf6b26627a93246c65cc3dc [formerly 93e92eb7974e1359abf6b26627a93246c65cc3dc [formerly 3c730fab2cae1bb52d34620af170a628b3b8c537]]
Former-commit-id: c6a853cf2ffb22d2656da34ec150620d229ff800
Former-commit-id: 6f655aac6ca8152b54ff10a360dfdfc04e32328f