branch: skip commit checks when deleting symref branches
Before a branch is deleted, we check that it points to a valid
commit.  With -d we also check that the commit is a merged; this
check is not done with -D.

The reason for that is that commits pointed to by branches should
never go missing; if they do then something broke and it's better
to stop instead of adding to the mess.  And a non-merged commit
may contain changes that are worth preserving, so we require the
stronger option -D instead of -d to get rid of them.

If a branch consists of a symref, these concerns don't apply.
Deleting such a branch can't make a commit become unreferenced,
so we don't need to check if it is merged, or even if it is
actually a valid commit.  Skip them in that case.  This allows
us to delete dangling symref branches.

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1f3c5b2d139bed75c226ae95fe78293dfbdf4102 [formerly 3c587996a03d7cd9e2d91e3334a87fad9075e5ae] [formerly 3c587996a03d7cd9e2d91e3334a87fad9075e5ae [formerly 0fe700e311f2d7e55eb23fe941fab9155d7f91df]]
Former-commit-id: 9824451fb37012d439a4d917958ee7fd2e7ac236
Former-commit-id: 40e36a3e1a9159af2eb2054577ae83e18d54d601