utf8: fix off-by-one wrapping of text
The wrapping logic in strbuf_add_wrapped_text() does currently not allow
lines that entirely fill the allowed width, instead it wraps the line one
character too early.

For example, the text "This is the sixth commit." formatted via
"%w(11,1,2)" (wrap at 11 characters, 1 char indent of first line, 2 char
indent of following lines) results in four lines: " This is", "  the",
"  sixth", "  commit." This is wrong, because "  the sixth" is exactly
11 characters long, and thus allowed.

Fix this by allowing the (width+1) character of a line to be a valid
wrapping point if it is a whitespace character.

Signed-off-by: Jan H. Schönherr <schnhrr@cs.tu-berlin.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: be71bd8ef326cf3379fa805afca0dcf02c72f9af [formerly be71bd8ef326cf3379fa805afca0dcf02c72f9af [formerly 14e1a4e1ff70aff36db3f5d2a8b806efd0134d50]]
Former-commit-id: c762bb1727e4d6b9002a9f8e4f3774888c9f01d1
Former-commit-id: 89119f29bb05160dfb11b065b5e65d071c95f60a