format-patch: make rfc2047 encoding more strict
RFC 2047 requires more characters to be encoded than it is currently done.
Especially, RFC 2047 distinguishes between allowed remaining characters
in encoded words in addresses (From, To, etc.) and other headers, such
as Subject.

Make add_rfc2047() and is_rfc2047_special() location dependent and include
all non-allowed characters to hopefully be RFC 2047 conformant.

This especially fixes a problem, where RFC 822 specials (e. g. ".") were
left unencoded in addresses, which was solved with a non-standard-conforming
workaround in the past (which is going to be removed in a follow-up patch).

Signed-off-by: Jan H. Schönherr <schnhrr@cs.tu-berlin.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 457336f11a9de297f3d73b6a4a0a7c2b8560fb46 [formerly 457336f11a9de297f3d73b6a4a0a7c2b8560fb46 [formerly 0fcec2ce5467740eb613da6a977deca3e7e866ba]]
Former-commit-id: 096068677301607800c0309de81a69abad5b9afa
Former-commit-id: 6dd6ab46015239d122279b1a746494174529dc3b