format-patch tests: check quoting/encoding in To: and Cc: headers
git-format-patch does currently not parse user supplied extra header
values (e. g., --cc, --add-header) and just replays them. That forces
users to add them RFC 2822/2047 conform in encoded form, e.g.

    --cc '=?UTF-8?q?Jan=20H=2E=20Sch=C3=B6nherr?= <...>'

which is inconvenient. We would want to update git-format-patch to
accept human-readable input

    --cc 'Jan H. Schönherr <...>'

and handle the encoding, wrapping and quoting internally in the future,
similar to what is already done in git-send-email. The necessary code
should mostly exist in the code paths that handle the From: and Subject:
headers.

Whether we want to do this only for the git-format-patch options
--to and --cc (and the corresponding config options) or also for
user supplied headers via --add-header, is open for discussion.

For now, add test_expect_failure tests for To: and Cc: headers as a
reminder and fix tests that would otherwise fail should this get
implemented.

Signed-off-by: Jan H. Schönherr <schnhrr@cs.tu-berlin.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 03a3b2974bd7ed1e7843f34d07123fb6d5c5c2cd [formerly 12ac93d23f955a0801343a319e41085073ef52c7] [formerly 12ac93d23f955a0801343a319e41085073ef52c7 [formerly 25dc8dad3a7c9f5426f549701e22eec1f894fec8]]
Former-commit-id: 4feabe9ae0c9462b575d2b297015a8c214532162
Former-commit-id: 098864f68e10434ba78fce0f51b13cc66afebbb9