Fix failure to delete a packed ref through a symref
When deleting a ref through a symref (e.g. using 'git update-ref -d HEAD'
to delete refs/heads/master), we would remove the loose ref, but a packed
version of the same ref would remain, the end result being that instead of
deleting refs/heads/master we would appear to reset it to its state as of
the last repack.

This patch fixes the issue, by making sure we pass the correct ref name
when invoking repack_without_ref() from within delete_ref().

Signed-off-by: Johan Herland <johan@herland.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4e110a3c97a5f0080d94b44913d9c27bc8e1d2e4 [formerly 4e110a3c97a5f0080d94b44913d9c27bc8e1d2e4 [formerly b274a7146c44624bc30bbd54124312d4d3e6a28d]]
Former-commit-id: 5c8065d909dec441c107c757bb038ca34ce4472f
Former-commit-id: d555cab46013ed518ffce66184cd53fb7f3e06d5