git symbolic-ref --delete $symref
Teach symbolic-ref to delete symrefs by adding the -d/--delete option to
git-symbolic-ref. Both proper and dangling symrefs are deleted by this
option, but other refs - or anything else that is not a symref - is not.

The symref deletion is performed by first verifying that we are given a
proper symref, and then invoking delete_ref() on it with the REF_NODEREF
flag.

Signed-off-by: Johan Herland <johan@herland.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 277f6e1f3eac7ee57d73f026bc4d3281757a5156 [formerly 277f6e1f3eac7ee57d73f026bc4d3281757a5156 [formerly 9ab55daa55377fa258c18ffb90211547e299eec9]]
Former-commit-id: 2bdeacbc43603798135675f1e91274cfdb1d3e38
Former-commit-id: 265b3e6d2267f2d11c63436a75c47424c99c85da