remote-testgit: properly check for errors
'feature done' was missing, which allowed fast-import exit properly, and
transport-helper to continue checking for refs and what not when in fact
the remote-helper died.

Let's enable that, and make sure the error paths are triggered.

Now transport-helper correctly detects the errors from fast-import,
unfortunately, not from fast-export because it might finish before
detecting a SIGPIPE. This means transport-helper will quit silently and
the user will not see any errors, which is bad. Hopefully the helper
will print the error before dying anyway, so not all is lost.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: b184803bd8fe7ab52dc539c0fd3f7b052ad8bb83 [formerly 0e00f8de9207d876d720685806636b90b53c9baa] [formerly 0e00f8de9207d876d720685806636b90b53c9baa [formerly 6c323322682abed3bb2a3904596b529ebccddb1c]]
Former-commit-id: 05ca270a2a00e4c237751d73794e0dfc126dfe25
Former-commit-id: 0d5182573644b8f38d09b511ed223fc843afd2e8