t1300: style updates
The t1300 test script is quite old, and does not use our
modern techniques or styles. This patch updates it in the
following ways:

  1. Use test_cmp instead of cmp (to make failures easier to
     debug).

  2. Use test_cmp instead of 'test $(command) = expected'.
     This makes failures much easier to debug, and also
     makes sure that $(command) exits appropriately.

  3. Use test_must_fail (easier to read, and checks more
     rigorously for signal death).

  4. Write tests with the usual style of:

       test_expect_success 'test name' '
               test commands &&
	       ...
       '

     rather than one-liners, or using backslash-continuation.
     This is purely a style fixup.

There are still a few command happening outside of
test_expect invocations, but they are all innoccuous system
commands like "cat" and "cp". In an ideal world, each test
would be self sufficient and all commands would happen
inside test_expect, but it is not immediately obvious how
the grouping should work (some of the commands impact the
subsequent tests, and some of them are setting up and
modifying state that many tests depend on). This patch just
picks the low-hanging style fruit, and we can do more fixes
on top later.

Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 8715af81f7716d8ca1c45c9c1637b45c14f12fed [formerly d242d5c3c9f0f56f79e2efae1b9447a6f05d393b] [formerly d242d5c3c9f0f56f79e2efae1b9447a6f05d393b [formerly ed838e661551958076beda19c9445101da2b70ae]]
Former-commit-id: 18d9048392c854a16584ed168f713a0d73486725
Former-commit-id: bb8baf514f1c0d0df91496fe078927664d53217e