git-config: fix regexp memory leaks on error conditions
The get_value function has a goto label for cleaning up on
errors, but it only cleans up half of what the function
might allocate. Let's also clean up the key and regexp
variables there.

Note that we need to take special care when compiling the
regex fails to clean it up ourselves, since it is in a
half-constructed state (we would want to free it, but not
regfree it).

Similarly, we fix git_config_parse_key to return NULL when
it fails, not a pointer to some already-freed memory.

Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 524da2e1fc52541e41891d0bdfbd9dea8fe9dcb5 [formerly 524da2e1fc52541e41891d0bdfbd9dea8fe9dcb5 [formerly 97ed50f93ba592b50278a8161282e862cb87b4c0]]
Former-commit-id: 80bd0a9aff410fb682eb277fe559a8d3662077f1
Former-commit-id: bb692fddfe192bd5cf5775fccad568cea8b4eb83