git-config: collect values instead of immediately printing
This is a refactor that will allow us to more easily tweak
the behavior for multi-valued variables, and it will
ultimately allow us to remove a lot git-config's custom code
in favor of the regular git_config code.

It does mean we're no longer streaming, and we're storing
more in memory for the --get-all case, but in practice it is
a tiny amount of data, and the results are instantaneous.

Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 76581d0d332a40044821592fe7dd7550e3a07092 [formerly 76581d0d332a40044821592fe7dd7550e3a07092 [formerly 7acdd6f0bc27bba37591e2d5fc2ca84e2af13ce4]]
Former-commit-id: 49f328ad047e2b657b058236a34421bad83ad738
Former-commit-id: 49261f30771a473c0cc6033e193406d409f6655d