git-pull: Avoid merge-base on detached head
git pull --rebase does some clever tricks to find the base
for $upstream, but it forgets that we may not have any
branch at all.  When this happens, git merge-base reports its
"usage" help in the middle of an otherwise successful
rebase operation, because git-merge is called with one too
few parameters.

Since we do not need the merge-base trick in the case of a
detached HEAD, detect this condition and bypass the clever
trick and the usage noise.

Signed-off-by: Phil Hord <hordp@cisco.com>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: e6e4e56dbf18977539832aad0d21b599b24cbd74 [formerly e6e4e56dbf18977539832aad0d21b599b24cbd74 [formerly e980765c59c173e9a10ce1069c763645b87fd50c]]
Former-commit-id: 530d3e307da16a6e75d54bbbfb684004dd5a0f69
Former-commit-id: 543eae807b2051602bb58b0ad78f05c17595b571