git-config: use git_config_with_options
The git-config command has always implemented its own file
lookup and parsing order. This was necessary because its
duplicate-entry handling did not match the way git's
internal callbacks worked. Now that this is no longer the
case, we are free to reuse the existing parsing code.

This saves us a few lines of code, but most importantly, it
means that the logic for which files are examined is
contained only in one place and cannot diverge.

Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 53129d7805fd21fa322f34e30d363ad4694ad5e6 [formerly f1bd6325cd3ba11fa40cde03eacd0b1029808ba1] [formerly f1bd6325cd3ba11fa40cde03eacd0b1029808ba1 [formerly e89558988305c3e912c0674faeb814e44563f7fd]]
Former-commit-id: 28c59b83a51c7fcf3349ac15ed78a23c75e5c315
Former-commit-id: 642402004836321ed80ba53da0b4f4f236cab1a7