Move setup_diff_pager to libgit.a
This is used by diff-no-index.c, part of libgit.a while it stays in
builtin/diff.c. Move it to diff.c so that we won't get undefined
reference if a program that uses libgit.a happens to pull it in.

While at it, move check_pager from git.c to pager.c. It makes more
sense there and pager.c is also part of libgit.a

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: e0430f8e37380d9ec0192d64b2f12d2188a7dbf9 [formerly e0430f8e37380d9ec0192d64b2f12d2188a7dbf9 [formerly 4914c9629c046f7f5abf4109ad756040f9ebe2bf]]
Former-commit-id: b9d18d21a9ab7d6c0dda4ebe818e1b942a90979f
Former-commit-id: 6df0221abddaa5f31e32f9bf06884a0ed61bdfea