t7407: Fix recursive submodule test
A test in t7404-submodule-foreach purports to test that
the --cached flag is properly noticed by --recursive calls
to the foreach command as it descends into nested
submodules.  However, the test really does not perform this
test since the change it looks for is in a top-level
submodule handled by the first invocation of the command.
To properly test for the flag being passed to recursive
invocations, the change must be buried deeper in the
hierarchy.

Move the change one level deeper so it properly verifies
the recursive machinery of the 'git submodule status'
command.

Signed-off-by: Phil Hord <hordp@cisco.com>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 30bfbb11f65b24b8c84bcb16f68e2c1cff5be00c [formerly e6607b6a85d9f4d3193caa55203a9423389624dc] [formerly e6607b6a85d9f4d3193caa55203a9423389624dc [formerly 58ca9ad4d68951f5f04f7c2399be55590dcf6a59]]
Former-commit-id: 20a8ca744eb874491c4b8385178ba817e66d1d29
Former-commit-id: 4e7eab79cf53d935064c4fe099edebcec9e70411