Teach --recursive to submodule sync
The submodule sync command was somehow left out when
--recursive was added to the other submodule commands.

Teach sync to handle the --recursive switch by recursing
when we're in a submodule we are sync'ing.

Change the report during sync to show submodule-path
instead of submodule-name to be consistent with the other
submodule commands and to help recursed paths make sense.

Signed-off-by: Phil Hord <hordp@cisco.com>
Acked-By: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 280b9211dfb7c84b37f28ad8d499d780e253988e [formerly 280b9211dfb7c84b37f28ad8d499d780e253988e [formerly 82f49f294c5ee58aff33dfd469bbd5c0d083e314]]
Former-commit-id: c9aee261be337da6378c577447e095a9eedea5e1
Former-commit-id: 5a1a010006099ca4fb9f40a92caae002ffe2d12e