mailmap: avoid out-of-bounds memory access
AddressSanitizer (http://clang.llvm.org/docs/AddressSanitizer.html)
complains of a one-byte buffer underflow in parse_name_and_email() while
running the test suite. And indeed, if one of the lines in the mailmap
begins with '<', we dereference the address just before the beginning of
the buffer when looking for whitespace to remove, before checking that
we aren't going too far.

So reverse the order of the tests to make sure that we don't read
outside the buffer.

Signed-off-by: Romain Francoise <romain@orebokech.com>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: d7ecd41655473da293c8c4bd738296334c0b394c [formerly d7ecd41655473da293c8c4bd738296334c0b394c [formerly 3174bc5ccfea34717a8dac2175c1951239a4985d]]
Former-commit-id: 08c44a050162b0800e6348caf80cd6bc8210cf58
Former-commit-id: 5a8127958bc2a9ffd114fb30eada62deccf253b3