diff_grep: use textconv buffers for add/deleted files
If you use "-G" to grep a diff, we will apply a configured
textconv filter to the data before generating the diff.
However, if the diff is an addition or deletion, we do not
bother running the diff at all, and just look for the token
in the added (or removed) content. This works because we
know that the diff must contain every line of content.

However, while we used the textconv-derived buffers in the
regular diff, we accidentally passed the original unmodified
buffers to regexec when checking the added or removed
content. This could lead to an incorrect answer.

Worse, in some cases we might have a textconv buffer but no
original buffer (e.g., if we pulled the textconv data from
cache, or if we reused a working tree file when generating
it). In that case, we could actually feed NULL to regexec
and segfault.

Reported-by: Peter Oberndorfer <kumbayo84@arcor.de>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: a588cc91260cdb151fdf1c4be77dcee39dcbaa51 [formerly a588cc91260cdb151fdf1c4be77dcee39dcbaa51 [formerly b1c2f57db326c43ec286855b396fb16b4890cf2d]]
Former-commit-id: a9513c00be0b00dc8a9b576976213e8d4c9ff7d9
Former-commit-id: 8860416fe51251db44c1b39eff62ad453dbb2b9a