pickaxe: use textconv for -S counting
We currently just look at raw blob data when using "-S" to
pickaxe. This is mostly historical, as pickaxe predates the
textconv feature. If the user has bothered to define a
textconv filter, it is more likely that their search string will be
on the textconv output, as that is what they will see in the
diff (and we do not even provide a mechanism for them to
search for binary needles that contain NUL characters).

This patch teaches "-S" to use textconv, just as we
already do for "-G".

Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 39a6add514e52a9a39c03ad28d537eb05457f3b2 [formerly 39a6add514e52a9a39c03ad28d537eb05457f3b2 [formerly ef90ab66e8eaa863777a8d383c5af2ff43a426d2]]
Former-commit-id: e731bbdbcaae8a222bb05d00cbadbd990c56efa9
Former-commit-id: 8654134171150f78125caa00c186942e39db0607