pickaxe: hoist empty needle check
If we are given an empty pickaxe needle like "git log -S ''",
it is impossible for us to find anything (because no matter
what the content, the count will always be 0). We currently
check this at the lowest level of contains(). Let's hoist
the logic much earlier to has_changes(), so that it is
simpler to return our answer before loading any blob data.

Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 98aeea126966123f08655fbd13f80cce2cf7656c [formerly 13e937818b7592d95b49c578d649aa5668734753] [formerly 13e937818b7592d95b49c578d649aa5668734753 [formerly 8fa4b09fb14339c5d62e65a88fde8b220cb1591b]]
Former-commit-id: 291db98f060bf1ce7f69fac003d62eb12dc26e6d
Former-commit-id: eee6aa1bf8c5fd9be709e6428ceb77a47c56e96c