string_list_longest_prefix(): remove function
This function was added in 60db085afb402436a91da7c597ee561b7b159fea [formerly f103f95b11d087f07c0c48bf784cd9197e18f203] in
the erroneous expectation that it would be used in the
reimplementation of longest_ancestor_length().  But it turned out to
be easier to use a function specialized for comparing path prefixes
(i.e., one that knows about slashes and root paths) than to prepare
the paths in such a way that a generic string prefix comparison
function can be used.  So delete string_list_longest_prefix() and its
documentation and test cases.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 059b37934c611b1b9b735e0310ba282a0c7f5eba