parse_dirstat_params(): use string_list to split comma-separated string
Use string_list_split_in_place() to split the comma-separated
parameters string.  This simplifies the code and also fixes a bug: the
old code made calls like

    memcmp(p, "lines", p_len)

which needn't work if p_len is different than the length of the
constant string (and could illegally access memory if p_len is larger
than the length of the constant string).

When p_len was less than the length of the constant string, the old
code would have allowed some abbreviations to be accepted (e.g., "cha"
for "changes") but this seems to have been a bug rather than a
feature, because (1) it is not documented; (2) no attempt was made to
handle ambiguous abbreviations, like "c" for "changes" vs
"cumulative".

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: e122fc823432abf31391187ae5a709ff1805fec2 [formerly e122fc823432abf31391187ae5a709ff1805fec2 [formerly 02e8ca0e50abc15596db1fa93bda29bbd91007a7]]
Former-commit-id: e1f0b5e3fdc4c8250ab6346eb87646559aa9c029
Former-commit-id: 7ca615ac5ae316c039a184723cae4ea4ec33c29a