submodule status: remove unused orig_* variables
When renaming orig_args to orig_flags in 98dbe63d (submodule: only
preserve flags across recursive status/update invocations) the call site
of the recursive cmd_status was forgotten. At that place orig_args is
still passed into the recursion, which is always empty since then. This
did not break anything because the orig_flags logic is not needed at all
when a function from the submodule script is called with eval, as that
inherits all the variables set by the option parsing done in the first
level of the recursion.

Now that we know that orig_flags and orig_args aren't needed at all,
let's just remove them from cmd_status().

Thanks-to: Phil Hord <hordp@cisco.com>
Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: eaf72716e1f68fbdb12401a3b12b9837b195a094 [formerly eaf72716e1f68fbdb12401a3b12b9837b195a094 [formerly e15bec0ec3b763449e418c12a5f2e128560c556d]]
Former-commit-id: 659202631014e4972e14b6307567865f1cdc059d
Former-commit-id: 3a1c0b14e84378ee878176a6d1a3067184412048