remote-curl: hoist gzip buffer size to top of post_rpc
When we gzip the post data for a smart-http rpc request, we
compute the gzip body and its size inside the "use_gzip"
conditional. We keep track of the body after the conditional
ends, but not the size. Let's remember both, which will
enable us to retry failed gzip requests in a future patch.

Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 7554b81f54e217773fdb76abe1fab65def05bb65 [formerly 7554b81f54e217773fdb76abe1fab65def05bb65 [formerly df126e108b899da133a980e900df39dfe57fcd59]]
Former-commit-id: 30c58e777c2b1d16a6d71ebf2b04cd2808798c1a
Former-commit-id: fb9dffe9125b58ba768945f773603ad847426833