strbuf_split_buf(): use ALLOC_GROW()
Use ALLOC_GROW() rather than inline code to manage memory in
strbuf_split_buf().  Rename "pos" to "nr" because it better describes
the use of the variable and it better conforms to the "ALLOC_GROW"
idiom.

Also, instead of adding a sentinal NULL value after each entry is
added to the list, only add it once after all of the entries have been
added.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: d3a3f60efa1690ccd42bc927578a10e69cf13373 [formerly d3a3f60efa1690ccd42bc927578a10e69cf13373 [formerly b8c2c1fa35f240ea8eee719c0f5a657285864573]]
Former-commit-id: 7d17d65706728a6917f50bad615241a075262c5f
Former-commit-id: 1decb77ae3b34928e9baa0a2dda227f806e4da02