strbuf_split_buf(): simplify iteration
While iterating, update str and slen to keep track of the part of the
string that hasn't been processed yet rather than computing things
relative to the start of the original string.  This eliminates one
local variable, reduces the scope of another, and reduces the amount
of arithmetic needed within the loop.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 0d26109bff790107a1940704acc86be43c64e889 [formerly 0d26109bff790107a1940704acc86be43c64e889 [formerly 1173bb331103b7e6d9e95549c7b7be12546d0697]]
Former-commit-id: 9fa9cacf46f111ca3d4c49a24e43c2f09af7bfab
Former-commit-id: 8c13132e2bd6805b97d1e8f83e3fa7ad59364e92