strbuf_split*(): rename "delim" parameter to "terminator"
The word "delimiter" suggests that the argument separates the
substrings, whereas in fact (1) the delimiter characters are included
in the output, and (2) if the input string ends with the delimiter,
then the output does not include a final empty string.  So rename the
"delim" arguments of the strbuf_split() family of functions to
"terminator", which is more suggestive of how it is used.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 8bcb34f8bacd5e164555730369128f762fd81f60 [formerly 8bcb34f8bacd5e164555730369128f762fd81f60 [formerly 17b73dc699c46d7af5d29d2f3813e7addafdce0d]]
Former-commit-id: 852bdc89de06c19dc55b0c3d3594edb94bff6192
Former-commit-id: 0ce94d9d2bfd26307d996bc0f6e27a8bc7a47819