gitweb.perl: fix %highlight_ext mappings
When commit 592ea41 refactored the list of extensions for
syntax highlighting, it failed to take into account perl's
operator precedence within lists. As a result, we end up
creating a dictionary of one-to-one elements when the intent
was to map mutliple related types to one main type (e.g.,
bash, ksh, zsh, and sh should all map to sh since they share
similar syntax, but we ended up just mapping "bash" to
"bash" and so forth).

This patch adds parentheses to make the mapping as the
original change intended. It also reorganizes the list to
keep mapped extensions together.

Signed-off-by: Richard Hubbell <richard_hubbe11@lavabit.com>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: ea58d56d20951e6efac5977d06d7454742ef4598 [formerly 10f2d9359b05ea2d140bfae2fdbb7d8545b90cb2] [formerly 10f2d9359b05ea2d140bfae2fdbb7d8545b90cb2 [formerly 048b3991924a0666a2ab4f381caac8d62962de38]]
Former-commit-id: 67b8109e752c297d36726d9c1f61a20570160c0f
Former-commit-id: 1bfa9aa019352c63a7668d5fceadeb8337a2873f