gitweb: make remote_heads config setting work
Git configuration items can not contain underscores in their section
and bottom-level variable name; the 'remote_heads' feature can not
be enabled on a per-repository basis with that name.

This changes the git-config option to be `gitweb.remoteheads` but does
not change the gitweb.conf option, to avoid backwards compatibility
issues.  We strip underscores from keys before looking through
git-config output for them.

An existing check on keynames was overly eager to reject non-word
letters, but if we ever start using three-level names, the middle
level string can contain almost anything, so fix that as well while
we are in the vicinity.

Signed-off-by: Phil Pennock <phil@apcera.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cb9aac48c93f009146e231747e7d556497e432d0 [formerly 1ce5f2f20f27839be9073de6f96f55a6944e0177] [formerly 1ce5f2f20f27839be9073de6f96f55a6944e0177 [formerly af507944a22a3dabe2041fe629b2178abb5c642f]]
Former-commit-id: 0792bf7b52af8450c42c772871226b39d8644420
Former-commit-id: 28fd5b93002d0d667120d40f10406033ed22e424