Update cygwin.c for new mingw-64 win32 api headers
The cygwin project recently switched to a new implementation of the
windows api, now using header files from the mingw-64 project. These
new header files are incompatible with the way cygwin.c included the
old headers: cygwin.c can be compiled using the new or the older (mingw)
headers, but different files must be included in different order for each
to work. The new headers are in use only for the current release series
(based upon the v1.7.x dll version). The previous release series using
the v1.5 dll is kept available but unmaintained for use on older versions
of Windows. So, patch cygwin.c to use the new include ordering only if
the dll version is 1.7 or higher.

Signed-off-by: Mark Levedahl <mlevedahl@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 7327fd305f2234aa7dae6b3eddaf88cf807a1137 [formerly 7327fd305f2234aa7dae6b3eddaf88cf807a1137 [formerly 380a4d927bff693c42fc6b22c3547bdcaac4bdc3]]
Former-commit-id: 57f75712570847da83ddba80689e5e41de78fca5
Former-commit-id: ccf69c82a382f324e9c4d2cf4b2cd89bdffb222a