diff: rename "set" variable
Once upon a time the builtin_diff function used one color, and the color
variables were called "set" and "reset". Nowadays it is a much longer
function and we use several colors (e.g., "add", "del"). Rename "set" to
"meta" to show that it is the color for showing diff meta-info (it still
does not indicate that it is a "color", but at least it matches the
scheme of the other color variables).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 920555b1504a739d0b623d55954a124c71767d85 [formerly 810c110959a15d24f5bd4ac47e4bbd8d68255927] [formerly 810c110959a15d24f5bd4ac47e4bbd8d68255927 [formerly d9c552f17a28e5ab1b72135038d4888ea02651ea]]
Former-commit-id: 63d80217615a27aa9f86daa2828eb04ded902139
Former-commit-id: 3c01ac628ce6934ab03ee970bb849abef4551f0c