test-lib: allow negation of prerequisites
You can set and test a prerequisite like this:

  test_set_prereq FOO
  test_have_prereq FOO && echo yes

You can negate the test in the shell like this:

  ! test_have_prereq && echo no

However, when you are using the automatic prerequisite
checking in test_expect_*, there is no opportunity to use
the shell negation.  This patch introduces the syntax "!FOO"
to indicate that the test should only run if a prerequisite
is not meant.

One alternative is to set an explicit negative prerequisite,
like:

  if system_has_foo; then
	  test_set_prereq FOO
  else
	  test_set_prereq NO_FOO
  fi

However, this doesn't work for lazy prerequisites, which
associate a single test with a single name. We could teach
the lazy prereq evaluator to set both forms, but the code
change ends up quite similar to this one (because we still
need to convert NO_FOO into FOO to find the correct lazy
script).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f6d2bdac80b49d8c90b9dc40bff5541c25ef4427 [formerly 7f4e859aa7c01ab0550276f28c5676b8e4d7b481] [formerly 7f4e859aa7c01ab0550276f28c5676b8e4d7b481 [formerly bdccd3c1fb261dc6d4aaf9fae446eea7136b76e2]]
Former-commit-id: fcfed9d26aa54fb88e6102e0aff8c64938fdbe9b
Former-commit-id: ea3ae867b75a8841890f2d774bcc321a1199a230