ident: keep separate "explicit" flags for author and committer
We keep track of whether the user ident was given to us
explicitly, or if we guessed at it from system parameters
like username and hostname. However, we kept only a single
variable. This covers the common cases (because the author
and committer will usually come from the same explicit
source), but can miss two cases:

  1. GIT_COMMITTER_* is set explicitly, but we fallback for
     GIT_AUTHOR. We claim the ident is explicit, even though
     the author is not.

  2. GIT_AUTHOR_* is set and we ask for author ident, but
     not committer ident. We will claim the ident is
     implicit, even though it is explicit.

This patch uses two variables instead of one, updates both
when we set the "fallback" values, and updates them
individually when we read from the environment.

Rather than keep user_ident_sufficiently_given as a
compatibility wrapper, we update the only two callers to
check the committer_ident, which matches their intent and
what was happening already.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f71fb0228f67fe5e1fdb1e5990aa4b18956d3b98 [formerly d1a6be1e69a42e62949398c06e8bac6097edde17] [formerly d1a6be1e69a42e62949398c06e8bac6097edde17 [formerly d6991ceedc6a6fbcc2184f8328eb1b4c88b55267]]
Former-commit-id: f62037a8e0ee06547b7afbad711d008cd6143e73
Former-commit-id: f1bdd61f1fec3a24792ce590214a8d505369ab77