checkout: print a message when switching unborn branches
When we switch to a new branch using checkout, we usually output a
message indicating what happened. However, when we switch from an unborn
branch to a new branch, we do not print anything, which may leave the
user wondering what happened.

The reason is that the unborn branch is a special case (see abe1998),
and does not follow the usual switch_branches code path. Let's add a
similar informational message to the special case to match the usual
code path.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 71b2af6bbcbdf33e5453b2a9eb4aabc7a0ad435f [formerly 71b2af6bbcbdf33e5453b2a9eb4aabc7a0ad435f [formerly afa8c07a26b9783c41b1066b75a35de1610a3cfa]]
Former-commit-id: 1988231ef5663baf1ecabe9f1aae3d5218ffb829
Former-commit-id: f59160ad36ea645ffece464f598e28901cdb7d1b