config: don't segfault when given --path with a missing value
When given a variable without a value, such as '[section] var' and
asking git-config to treat it as a path, git_config_pathname returns
an error and doesn't modify its output parameter. show_config assumes
that the call is always successful and sets a variable to indicate
that vptr should be freed. In case of an error however, trying to do
this will cause the program to be killed, as it's pointing to memory
in the stack.

Detect the error and return immediately to avoid freeing or accessing
the uninitialed memory in the stack.

Signed-off-by: Carlos Martín Nieto <cmn@elego.de>
Acked-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 889ca3efa683d49f6d821b0ed06bdb294c030849 [formerly e5b8120b0cefc9d9b1ef1034c2c56c5721fde514] [formerly e5b8120b0cefc9d9b1ef1034c2c56c5721fde514 [formerly 962c38eedd931a608b5e675f12a21569d9f4d39f]]
Former-commit-id: bce20cbcbd68e8cfaa1e5fa5bb7dd7282a56b7fd
Former-commit-id: 13f93a0c120c6ba48d7e4851d2a48cf33f2805a1