completion: add new zsh completion
It seems there's always issues with zsh's bash completion emulation.
I've tried to fix as many as I could[1], and most of the fixes are already
in the latest version of zsh, but still, there are issues.

There is no point going through all that pain; the emulation is easy to
achieve, and this patch works better than zsh's bash completion
emulation.

[1] http://zsh.git.sourceforge.net/git/gitweb.cgi?p=zsh/zsh;a=commitdiff;h=23907bb840c80eef99eabba17e086e44c9b2d3fc

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d8c9c3acc3def063888426ffcf61f545b5cd27bd [formerly 031ee34ff2b513945e74cf30486c07b8b4671f8b] [formerly 031ee34ff2b513945e74cf30486c07b8b4671f8b [formerly c9407860f47ec4cf1c64a2c2ca4ae0b451469658]]
Former-commit-id: c2006b53e36587df2024f6ad21ce2797c2d7f6b5
Former-commit-id: bacd1fb47c9801bf776b9727fae4ee7706ee6b69