Makefile: hide stderr of curl-config test
You will get

    $ make distclean 2>&1 | grep curl
    /bin/sh: curl-config: not found
    /bin/sh: curl-config: not found
    /bin/sh: curl-config: not found
    /bin/sh: curl-config: not found
    /bin/sh: curl-config: not found
    $

if you don't have a curl development package installed.

The intent is not to alarm the user, but just to test if there is
a new enough curl installed.  However, if you look at search engine
suggested completions, the above "error" messages are confusing
people into thinking curl is a hard requirement.

Redirect this error output to /dev/null as it is not necessary to be
shown to the end users.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2a7bffcad99286463121a63411d397ec9c367f1e [formerly 2a7bffcad99286463121a63411d397ec9c367f1e [formerly cad06d4d78b52b99d451063db2e6d08c34697d32]]
Former-commit-id: f0ba39496615de54475f02196c6631ddf8e46083
Former-commit-id: 68f5169e7048442ea7c4c3292c59035506b42b7b