git-send-email: remove garbage after email address
In some cases it is useful to add additional information after the
email address on the Cc: footer in a commit log, for instance:

"Cc: Stable kernel <stable@vger.kernel.org> #v3.4 v3.5 v3.6"

However, git-send-email refuses to pick up such an invalid address
when the Email::Valid perl module is available, or just uses the
whole line as the email address.

In sanitize_address(), remove everything after the email address, so
that the result is a valid email address that makes Email::Valid
happy.

Signed-off-by: Krzysztof Mazur <krzysiek@podlesie.net>
Tested-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5082c8f1d13b40d9fa636cec9b21103fa9cccc76 [formerly 146371a7196eeec48cf1e8e84f6e68ed70e9606d] [formerly 146371a7196eeec48cf1e8e84f6e68ed70e9606d [formerly 831a488b76e0fcad68c6d1c7aaf9f086199e1738]]
Former-commit-id: 32626296730823da2301261ffde30887ceac0ad5
Former-commit-id: 13b5d8a99798fb8a5eea784cd6345d9922e0b306