git-send-email: fix fallback code in extract_valid_address()
In the fallback check, used when Email::Valid is not available, the
extract_valid_address() uses $1 without checking for success of matching
regex. The $1 variable may still hold the result of previous match,
which is the address when email address was in '<>' or be undefined
otherwise.

Now if match fails undefined value is always returned to indicate error.
The same value is used by Email::Valid->address() in that case.

Previously 'foo@bar' address was rejected by Email::Valid and fallback,
but '<foo@bar>' was rejected by Email::Valid, but accepted by fallback.

Signed-off-by: Krzysztof Mazur <krzysiek@podlesie.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8727d67ff13f3be58c71e782299ad6139f4b1b25 [formerly 8727d67ff13f3be58c71e782299ad6139f4b1b25 [formerly 95c0d4b68a0fcba3adf45ce3a8f1edf99a9030a1]]
Former-commit-id: 15e9dfa3f9794785529dac62b52aa3a9098ce294
Former-commit-id: 63226c50776bcdcb0de80bdf492d321ec9136173