Teach rm to remove submodules when given with a trailing '/'
Doing a "git rm submod/" on a submodule results in an error:
	fatal: pathspec 'submod/' did not match any files
This is really inconvenient as e.g. using TAB completion in a shell on a
submodule automatically adds the trailing '/' when it completes the path
of the submodule directory. The user has then to remove the '/' herself to
make a "git rm" succeed. Doing a "git rm -r somedir/" is working fine, so
there is no reason why that shouldn't work for submodules too.

Teach git rm to not error out when a '/' is appended to the path of a
submodule. Achieve this by chopping off trailing slashes from the path
names given if they represent directories. Add tests to make sure that
logic only applies to directories and not to files.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 97a7960e4d19efd37d3c054493c548ba5fb69d2b [formerly 97a7960e4d19efd37d3c054493c548ba5fb69d2b [formerly 53e4c5dcab33d2718476867b865da9679bd2c24a]]
Former-commit-id: 5e64eff4e16a1dd35563fe439e480dfdbb3a1ba6
Former-commit-id: 5b2fbd486bd59d12ab61cd729666e19dd0a21a3a