git p4: catch p4 describe errors
Group the two calls to "p4 describe" into a new helper function,
and try to validate the p4 results.  The current behavior when p4
describe fails is to die with a python backtrace.  The new behavior
will print the full response.

This does not solve any particular problem, but adds more
checking in hopes of narrowing down odd behavior seen on
at least two occasions.

Based-on-patch-by: Matt Arsenault <arsenm2@gmail.com>
Reported-by: Arthur <a.foulon@amesys.fr>
Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f944b4940575739a7f79625a7545cdf80ee941a9 [formerly f944b4940575739a7f79625a7545cdf80ee941a9 [formerly 18fa13d0b34b6243d3679ea78325ee33ee4d0989]]
Former-commit-id: 5337ebe2fcb96b38d807e809457d3aac9082ec62
Former-commit-id: 25e6301d9c57d69f6687dcad2599cdeee83e3102