git p4: handle servers without move support
Support for the "p4 move" command was added in 8e9497c (git p4:
add support for 'p4 move' in P4Submit, 2012-07-12), which checks
to make sure that the client and server support the command.

But older versions of p4d may not handle the "-k" argument, and
newer p4d allow disabling "p4 move" with a configuration setting.
Check for both these cases by testing a p4 move command on bogus
filenames and looking for strings in the error messages.

Reported-by: Vitor Antunes <vitor.hda@gmail.com>
Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 30b56bc58ce4a800f882ee8895fe0591a350d6bc [formerly 5d2a4bed8dc7fcceb428d9e50d1dc7b83531af7e] [formerly 5d2a4bed8dc7fcceb428d9e50d1dc7b83531af7e [formerly 249da4c0dcd0534f416e2d5da0a9923c6068e492]]
Former-commit-id: eaf4e08abf97ae7597fdb8118c30c2ab21e4fc67
Former-commit-id: 9a2903cf59b4f9a3d7ca578101d6811d7257a7a3