fast-export: avoid importing blob marks
We want to be able to import, and then export, using the same marks, so
that we don't push things that the other side already received.

Unfortunately, fast-export doesn't store blobs in the marks, but
fast-import does. This creates a mismatch when fast export is reusing a
mark that was previously stored by fast-import.

There is no point in one tool saving blobs, and the other not, but for
now let's just check in fast-export that the objects are indeed commits.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4abae6314c904a6cf5da55f099e61f5b4a528c34 [formerly 0cec6dcde10eee572f927ac74b6163384f8bc81c] [formerly 0cec6dcde10eee572f927ac74b6163384f8bc81c [formerly 5d3698ffb4ccf8b842814559007e011fc409650b]]
Former-commit-id: 27838e606f112ebd168d6d51451ec3fbbc140b2f
Former-commit-id: 9a9c387326cd80d39a403f0b1d11952ee4f42665