lf_to_crlf(): NUL-terminate msg_data::data
Through the rest of the file, the data member of struct msg_data is
kept NUL-terminated, and that fact is relied upon in a couple of
places.  Change lf_to_crlf() to preserve this invariant.

In fact, there are no execution paths in which lf_to_crlf() is called
and then its data member is required to be NUL-terminated, but it is
better to be consistent to prevent future confusion.

Document the invariant in the struct msg_data definition.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: eb29dcdcab692000bd1bb1321137e465e493c604 [formerly fb3949535af57afa67059cc0bff63b7d30ff756f] [formerly fb3949535af57afa67059cc0bff63b7d30ff756f [formerly 32a8569ecf211cc2ba98333daa10b1a6e783ae87]]
Former-commit-id: 2140895d57ba2c99ad3387ab3fb36cc7d749f999
Former-commit-id: 367e15228210d2ab23e95322ce873e84c546f063