imap-send: correctly report errors reading from stdin
Previously, read_message() didn't distinguish between an error and eof
when reading its input.  This could have resulted in incorrect
behavior if there was an error: (1) reporting "nothing to send" if no
bytes were read or (2) sending an incomplete message if some bytes
were read before the error.

Change read_message() to return -1 on ferror()s and 0 on success, so
that the caller can recognize that an error occurred.  (The return
value used to be the length of the input read, which was redundant
because that is already available as the strbuf length.

Change the caller to report errors correctly.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 98872f09e87e38b111beddb940039d6e2ccdb3c7 [formerly 98872f09e87e38b111beddb940039d6e2ccdb3c7 [formerly 6360bee4cd902bdea5db5826821edffd6f367d89]]
Former-commit-id: 2e4ea556c6030d62daef2825ea060b7b15f827e1
Former-commit-id: ddad752606cfc0ecaf4b774626d26218504a547b