imap-send: change msg_data from storing (ptr, len) to storing strbuf
struct msg_data stored (ptr, len) of the data to be included in a
message, kept the character data NUL-terminated, etc., much like a
strbuf would do.  So change it to use a struct strbuf.  This makes
the code clearer and reduces copying a little bit.

A side effect of this change is that the memory for each message is
freed after it is used rather than leaked, though that detail is
unimportant given that imap-send is a top-level command.

By the way, there is a bunch of infrastructure in this file for
dealing with IMAP flags, although there is nothing in the code that
actually allows any flags to be set.  If there is no plan to add
support for flags in the future, a bunch of code could be ripped out
and "struct msg_data" could be completely replaced with strbuf, but
that would be a separate topic.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 09743a84f0172fd6bafb6a9276b94d979652dda3 [formerly 09743a84f0172fd6bafb6a9276b94d979652dda3 [formerly f035ab620520dbf286d3303194f10a489d7c7f56]]
Former-commit-id: bb172e5c5804148972af88b9dd20c45ad090623a
Former-commit-id: 65954062621a7ecd3d419fd1664b0155eb1fe5ae