Support for git aliasing for tcsh completion
tcsh users sometimes alias the 'git' command to another name.  In
this case, the user expects to only have to issue a new 'complete'
command using the alias name.

However, the tcsh script currently uses the command typed by the
user to call the appropriate function in git-completion.bash, either
_git() or _gitk().  When using an alias, this technique no longer
works.

This change specifies the real name of the command (either 'git' or
'gitk') as a parameter to the script handling tcsh completion.  This
allows the user to use any alias for the 'git' or 'gitk' commands,
while still getting completion to work.

A check for the presence of ${HOME}/.git-completion.bash is also
added to help the user make use of the script properly.

Signed-off-by: Marc Khouzam <marc.khouzam@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6aec2a0b691082d374213e496bdea25cfefd724c [formerly 6aec2a0b691082d374213e496bdea25cfefd724c [formerly ce45ea6a133a68ead5d06b33453cc5296fb5c720]]
Former-commit-id: 68ec109275d7cbf8e1430e7c3d39b9be3b2c7a6c
Former-commit-id: 6c9804a40f429afa088e8625abdc10978d90697f