diff --stat: do not count "unmerged" entries
Even though we show a separate *UNMERGED* entry in the patch and
diffstat output (or in the --raw format, for that matter) in
addition to and separately from the diff against the specified stage
(defaulting to #2) for unmerged paths, they should not be counted in
the total number of files affected---that would lead to counting the
same path twice.

The separation done by the previous step makes this fix simple and
straightforward.  Among the filepairs in diff_queue, paths that
weren't modified, and the extra "unmerged" entries do not count as
total number of files.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2d9b095fba2985f8591e4df6f18b9608edcb4a46 [formerly 2d9b095fba2985f8591e4df6f18b9608edcb4a46 [formerly 82dfc2c44ecda8a7afe417086c704b141a11cd58]]
Former-commit-id: f695adf8fa9dfe1f53e740faaf7be8e6cdfbefff
Former-commit-id: 34dec088b1d4b97ecb2f5d88f3ad7a7b3f32ff7e