git-remote-helpers.txt: document invocation before input format
In the distant past, the order things were documented was
'Invocation', 'Commands', 'Capabilities', ...

Then it was decided that before giving a list of Commands, there
should be an overall description of the 'Input format', which was
a wise decision. However, this description was put as the very
first thing, with the rationale that any implementor would want
to know that first.

However, it seems an implementor would actually first need to
know how the remote helper will be invoked, so moving
'Invocation' to the front again seems logical. Moreover, we now
don't switch from discussing the input format to the invocation
style and then back to input related stuff.

Signed-off-by: Max Horn <max@quendi.de>
Acked-by: Sverre Rabbelier <srabbelier@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c536055478f5c71c4b30266380fe83893e74871b [formerly c536055478f5c71c4b30266380fe83893e74871b [formerly f829a9eae6adc7e1d02fa28787e43fccaf98c0cc]]
Former-commit-id: fae89c92cd9c30179068c165cb5c3ce09fb4a2a5
Former-commit-id: 55c3ac45fa6d9d8a41bae7365ce060d1faece2bb