git-remote-helpers.txt: document invocation before input format
In the distant past, the order things were documented was
'Invocation', 'Commands', 'Capabilities', ...

Then it was decided that before giving a list of Commands, there
should be an overall description of the 'Input format', which was
a wise decision. However, this description was put as the very
first thing, with the rationale that any implementor would want
to know that first.

However, it seems an implementor would actually first need to
know how the remote helper will be invoked, so moving
'Invocation' to the front again seems logical. Moreover, we now
don't switch from discussing the input format to the invocation
style and then back to input related stuff.

Signed-off-by: Max Horn <max@quendi.de>
Acked-by: Sverre Rabbelier <srabbelier@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bbe728ff4d697fb9f74c1b8684bcfc36ef7631c5 [formerly 4db58ebcbfd298f5223550a5211b45c75aa426cc] [formerly 4db58ebcbfd298f5223550a5211b45c75aa426cc [formerly 6cf9614df676dec3b5c14ecc98f917e450a80fef]]
Former-commit-id: a373984099ffa4991562e5ffe58a05efa0b2768a
Former-commit-id: ea26d7a280fa383ba8309461e9e2940c97abbdb3