git-remote-helpers.txt: clarify options & ref list attributes
The documentation was misleading in that it gave the impression that
'for-push' could be used as a ref attribute in the output of the
'list' command. That is wrong.

Also, explicitly point out the connection between the commands
'list' and 'options' on the one hand, and the sections
'REF LIST ATTRIBUTES' and 'OPTIONS' on the other hand.

Signed-off-by: Max Horn <max@quendi.de>
Acked-by: Sverre Rabbelier <srabbelier@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 53492fe8f34aa5f54873dfd0c556c68d0e82f44e [formerly 2eb5068dd560ea6a5e0a65682175e786913422d7] [formerly 2eb5068dd560ea6a5e0a65682175e786913422d7 [formerly 9ec8bcda6031d2c75fb8b06b26bcf9414e9448f6]]
Former-commit-id: b0bcd9c1e21be369d8c4b0db42ed81395ab03f2c
Former-commit-id: b2ae22f81666d4faa71cdf71ba290f7afd49fe7a