configure.ac: fix pthreads detection on Mac OS X
The configure script checks whether certain flags are required to use
pthreads. But it did not consider that *none* might be needed (as is the
case on Mac OS X). This lead to configure adding "-mt" to the list of
flags (which does nothing on OS X except producing a warning). This in
turn triggered a compiler warning on every single file.

To solve this, we now first check if pthreads work without extra flags.
This means the check is now order dependant, hence a comment is added
explaining this, and the reasons for it.

Note that it might be possible to write an order independent test, but
it does not seem worth the extra effort required for implementing and
testing such a solution, when this simple solution exists and works.

Signed-off-by: Max Horn <max@quendi.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8e0be10b6c2fae5d7f073c25a582bd6dfdf88f4f [formerly 8e0be10b6c2fae5d7f073c25a582bd6dfdf88f4f [formerly e0a52279304a67d6e13c5a5b4165dd09f6ba61aa]]
Former-commit-id: 8029fa4c3d7d23e41398a51193da715518dcbc5a
Former-commit-id: 8a6e728a250afcb0a4bdc65e45d8ac01ffac31ce