fsck: warn about '.' and '..' in trees
A tree with meta-paths like '.' or '..' does not work well
with git; the index will refuse to load it or check it out
to the filesystem (and even if we did not have that safety,
it would look like we were overwriting an untracked
directory). For the same reason, it is difficult to create
such a tree with regular git.

Let's warn about these dubious entries during fsck, just in
case somebody has created a bogus tree (and this also lets
us prevent them from propagating when transfer.fsckObjects
is set).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b9497a57d07556a37400f858abd6fcb8e79f1fda [formerly b9497a57d07556a37400f858abd6fcb8e79f1fda [formerly 5d34a4359d51b44483a1e62dc8a1f6ec686c8b6f]]
Former-commit-id: 9ae7613d461a344181e0756c903b5c8795be3078
Former-commit-id: 62ae60e0001230005a34bab07378f67d15b10585