fsck: warn about ".git" in trees
Having a ".git" entry inside a tree can cause confusing
results on checkout. At the top-level, you could not
checkout such a tree, as it would complain about overwriting
the real ".git" directory. In a subdirectory, you might
check it out, but performing operations in the subdirectory
would confusingly consider the in-tree ".git" directory as
the repository.

The regular git tools already make it hard to accidentally
add such an entry to a tree, and do not allow such entries
to enter the index at all. Teaching fsck about it provides
an additional safety check, and let's us avoid propagating
any such bogosity when transfer.fsckObjects is on.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 61f123887c2ce2a53bc04b7190a1f6db3d55929a [formerly 381727251d8914d2703a4688fb51885407b62c0b] [formerly 381727251d8914d2703a4688fb51885407b62c0b [formerly 5c17f51270e1b384e03fb9c16b5a0040b115ae8c]]
Former-commit-id: 4cd562c86ee157d135a9bafac56dc499b5058464
Former-commit-id: 9a6305197700e2edfa7998d1cdbe97f643a619f1