Rename git-remote-testgit to git-remote-testpy
This script is not really exercising the remote-helper functionality,
but more the python framework for remote helpers that live in
git_remote_helpers.

It's also not a good example of how to write remote-helpers, unless you
are planning to use python, and even then you might not want to use this
framework.

So let's use a more appropriate name: git-remote-testpy.

A patch that replaces git-remote-testgit with a simpler version is on
the way.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 955d09d823aeea7de3d12e51ab31f4f567149451 [formerly 955d09d823aeea7de3d12e51ab31f4f567149451 [formerly d0ac3ffd9da8d1c3e629129030f74e1927faf575]]
Former-commit-id: 9bf0b1f4ece47c0a4ec07f3ebafc5acbb5c0914b
Former-commit-id: 34a07c9c13500e9c0a3e9466fbd513ade841ed61