Add new simplified git-remote-testgit
Exercising the python remote helper framework is for another tool and
another test. This is about testing the remote-helper interface.

It's way simpler, it exercises the same features of remote helpers, it's
easy to read and understand, and it doesn't depend on python.

For now let's just copy the old remote-helpers test script, although
some of those tests don't make sense. In addition, this script would be
able to test other features not currently being tested.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fcb7b1cd9a28682ee1b9889ac1443c05be195ab2 [formerly fcb7b1cd9a28682ee1b9889ac1443c05be195ab2 [formerly fc407f9821840041757975045c4a1ca01effa27a]]
Former-commit-id: 66010856dd9dfff542ddfd2f4d1d49568768a84a
Former-commit-id: fc9ec12d1b9ae715e762a3aac495cea03523e3b9