remote-testgit: implement the "done" feature manually
People who want to write their own remote-helper will find it more
useful to see clearly how they are supposed to advertise and implement
the "done" feature themselves.

Right now we are relying on fast-export to do that by using the
--use-done-feature argument. However, people writing their own
remote-helper would probably not have such an option, as they would
probably be writing the fast-export functionality themselves.

It should now be clearer to them.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fcb84850ecd5114ffc7f43590bb2d299583f9706 [formerly fcb84850ecd5114ffc7f43590bb2d299583f9706 [formerly 1d3f9a30938e12b8ef7b807df7565b40b743d421]]
Former-commit-id: 9c78b1fcce00d95e9a7c174737219f77736bb15b
Former-commit-id: 847043807d320496c6dee406453f5cb7e359c3e9