t4049: refocus tests
The primary thing Linus's patch wanted to change was to make sure
that 0-line change appears for a mode-only change.  Update the
first test to chmod a file that we can see in the output (limited
by --stat-count) to demonstrate it.  Also make sure to use test_chmod
and compare the index and the tree, so that we can run this test
even on a filesystem without permission bits.

Later two tests are about fixes to separate issues that were
introduced and/or uncovered by Linus's patch as a side effect, but
the issues are not related to mode-only changes.  Remove chmod from
the tests.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6f6504565803b9b1c5cd460ab760d7b15f320c7c [formerly 6f6504565803b9b1c5cd460ab760d7b15f320c7c [formerly de9095955c581a664d9ce568a1ae58a98484090e]]
Former-commit-id: b67cc5fcd34118085696b85d0daac14ba6af3384
Former-commit-id: 8a240ad80eadb225513305297a90a4cbe6edec90