git-fast-import.txt: improve documentation for quoted paths
The documentation mentioned only newlines and double quotes as
characters needing escaping, but the backslash also needs it. Also, the
documentation was not clearly saying that double quotes around the file
name were required (double quotes in the examples could be interpreted as
part of the sentence, not part of the actual string).

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 79570baecb49fa5c4ec48722c3bac6ef6b025e37 [formerly 79570baecb49fa5c4ec48722c3bac6ef6b025e37 [formerly 7c65b2ebb72fcf9b563be3367a088256757343a6]]
Former-commit-id: 9376e3c46ca06fde3ba6a520e375bfc0007b09c1
Former-commit-id: 9f19b3e3a8513b4efceb21f8ea06e68df5825e15