push: require force for refs under refs/tags/
References are allowed to update from one commit-ish to another if the
former is an ancestor of the latter.  This behavior is oriented to
branches which are expected to move with commits.  Tag references are
expected to be static in a repository, though, thus an update to
something under refs/tags/ should be rejected unless the update is
forced.

Signed-off-by: Chris Rorvick <chris@rorvick.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4131848fcedef1253364d828679b3fc4881ae82c [formerly 4131848fcedef1253364d828679b3fc4881ae82c [formerly dbfeddb12e5bb540ed3c852eebda3df9117bd150]]
Former-commit-id: fef376b0c9c55be241977ef4f6d7b3d515eba366
Former-commit-id: e3b0930e1c7b26e4e99bc564e1adfbc39e6be6d0