push: clarify rejection of update to non-commit-ish
Pushes must already (by default) update to a commit-ish due to the fast-
forward check in set_ref_status_for_push().  But rejecting for not being
a fast-forward suggests the situation can be resolved with a merge.
Flag these updates (i.e., to a blob or a tree) as not forwardable so the
user is presented with more appropriate advice.

While updating *from* a tag object is potentially destructive, updating
*to* a tag is not.  Additionally, a push to the refs/tags/ hierarchy is
already excluded from fast-forwarding, and refs/heads/ is protected from
anything but commit objects by a check in write_ref_sha1().  Thus
someone fast-forwarding to a tag is probably not doing so by accident.
Since updating to a tag is benign and unlikely to cause confusion, allow
it in case someone finds the behavior useful.

Signed-off-by: Chris Rorvick <chris@rorvick.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 15e61a102deb3a02e1fe368689db2a8ac0eab31a [formerly 15e61a102deb3a02e1fe368689db2a8ac0eab31a [formerly 80054cf9d53aad76631797d76ce33a56c855c61a]]
Former-commit-id: 9a1a33b524479ba66397643e5a3356d9fa17c739
Former-commit-id: e00b49ab02b39ff68b644dc5fc6791a7cf6b2f28