launch_editor: ignore terminal signals while editor has control
The user's editor likely catches SIGINT (ctrl-C).  but if
the user spawns a command from the editor and uses ctrl-C to
kill that command, the SIGINT will likely also kill git
itself (depending on the editor, this can leave the terminal
in an unusable state).

Let's ignore it while the editor is running, and do the same
for SIGQUIT, which many editors also ignore. This matches
the behavior if we were to use system(3) instead of
run-command.

Signed-off-by: Paul Fox <pgf@foxharp.boston.ma.us>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d2434171f8682b6774ac8c6e4c1a2ef9165320d7 [formerly d2434171f8682b6774ac8c6e4c1a2ef9165320d7 [formerly 913ef36093eac3ec78b5fb155cc2beb5843b1ce5]]
Former-commit-id: a7d773a8802abec4185475e0131c883896c12262
Former-commit-id: 64ebd5405d19d85024373b4e4c268410a61052ab