mingw: reuse tty-version of git_terminal_prompt
The getpass-implementation we use on Windows isn't at all ideal;
it works in raw-mode (as opposed to cooked mode), and as a result
does not deal correcly with deletion, arrow-keys etc.

Instead, use cooked mode to read a line at the time, allowing the
C run-time to process the input properly.

Since we set files to be opened in binary-mode by default on
Windows, introduce a FORCE_TEXT macro that expands to the "t"
modifier that forces the terminal to be opened in text-mode so we
do not have to deal with CRLF issues.

Signed-off-by: Erik Faye-Lund <kusmabite@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e43e6e8ba2902dfd026149f4bc8c7b02957a4b3b [formerly e43e6e8ba2902dfd026149f4bc8c7b02957a4b3b [formerly afb43561b81b435216b5b8c59fdaa48255e662fb]]
Former-commit-id: 2495f054a94e76344edd4a2cea417b50a6e87146
Former-commit-id: 683da0dfa62337353c1da0a4b4aa1e500927bb13