merge-tree: add comments to clarify what these functions are doing
Rename the "branch1" parameter given to resolve() to "ours", to
clarify what is going on.  Also, annotate the unresolved_directory()
function with some comments to show what decisions are made in each
step, and highlight two bugs that need to be fixed.

Add two tests to t4300 to illustrate these bugs.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6ec422359d9ffb0d6764d2ff451b0031dd881c59 [formerly 6ec422359d9ffb0d6764d2ff451b0031dd881c59 [formerly 8dd15c6a909c059c6fd4f4dd914dbf932617ea57]]
Former-commit-id: e04ca54291d6534ddf4f30be216fd21be359f2be
Former-commit-id: 617de384d2b41ee753610c23fdd14ef840a862b9