sh-setup: work around "unset IFS" bug in some shells
With an unset IFS, field splitting is supposed to act as if IFS is
set to the usual SP HT LF, but Marc Branchaud reports that the shell
on FreeBSD 7.2 gets this wrong.

It is easy to set it to the default value manually, and it is also
safer in case somebody tries to save the old value away and restore,
e.g.

	$oIFS=$IFS
	IFS=something
	...
	IFS=$oIFS

while forgetting that the original IFS might be unset (which can be
coded but would be more involved).

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1d5982ab4878fcf135bd094589fbac5cb9246f38 [formerly 1d5982ab4878fcf135bd094589fbac5cb9246f38 [formerly 393050c32b58dceef8706d389abe3793fbac51ae]]
Former-commit-id: 0d01154786f78d79db10528bd57b6c4728422548
Former-commit-id: b2988e68961e773d34c51f2bfe7cdd5ca7818bcb