mingw_rmdir: do not prompt for retry when non-empty
in ab1a11be ("mingw_rmdir: set errno=ENOTEMPTY when appropriate"),
a check was added to prevent us from retrying to delete a directory
that is both in use and non-empty.

However, this logic was slightly flawed; since we didn't return
immediately, we end up falling out of the retry-loop, but right into
the prompting-loop.

Fix this by setting errno, and guarding the prompting-loop with an
errno-check.

Signed-off-by: Erik Faye-Lund <kusmabite@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 572e2481fe9d1852a4ecd3035772d6ac38cbbb33 [formerly 572e2481fe9d1852a4ecd3035772d6ac38cbbb33 [formerly a83b2b578c04a2abe33a586ec9c64e75ef5bc819]]
Former-commit-id: 86890c41699ceb3d9b0acc51f4c586039c29e742
Former-commit-id: 62a6f80f2e0dc294f47548117ef2e48162e9ed5b