shortlog: fix wrapping lines of wraplen
A recent commit [1] fixed a off-by-one wrapping error.  As a
side-effect, the conditional in add_wrapped_shortlog_msg() to decide
whether to append a newline needs to be removed.  The function
should always append a newline, which was the case before the
off-by-one fix, because strbuf_add_wrapped_text() never returns a
value of wraplen; when it returns wraplen, the string does not end
with a newline, so this caller needs to add one anyway.

[1] 1ddb7b32ad7f52c14683e0c75e5cfa4e893a6914 [formerly 89119f29bb05160dfb11b065b5e65d071c95f60a] [formerly c762bb1727e4d6b9002a9f8e4f3774888c9f01d1] [formerly be71bd8ef326cf3379fa805afca0dcf02c72f9af [formerly be71bd8ef326cf3379fa805afca0dcf02c72f9af [formerly 14e1a4e1ff70aff36db3f5d2a8b806efd0134d50]]] utf8: fix off-by-one
    wrapping of text

Signed-off-by: Steffen Prohaska <prohaska@zib.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ce5c3edcb7ab425a8e72a4a3efc086fe73e149e5 [formerly ce5c3edcb7ab425a8e72a4a3efc086fe73e149e5 [formerly 5b59708268c82262fb13fb8681663dd158b030bd]]
Former-commit-id: 3ab7b29baea5ba60f4a4a74f1118d8482c0dbdad
Former-commit-id: b40fadaab1d1f7afbba6f5cd410f26964c9ba5b2