gitweb: Sort projects with undefined ages last
Sorting gitweb's project list by age ('Last Change') currently shows
projects with undefined ages at the head of the list. This gives a less
useful result when there are a number of projects that are missing or
otherwise faulty and one is trying to see what projects have been
updated recently.

Fix by sorting these projects with undefined ages at the bottom of the
list when sorting by age.

Signed-off-by: Matthew Daley <mattjd@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b2add59ddac2692458f38cc592f9ba908aa6a177 [formerly b2add59ddac2692458f38cc592f9ba908aa6a177 [formerly 28dae1812bc70f06079cb7d3fd4cdbfa9e59cbeb]]
Former-commit-id: 8cea2f67a6dec01cfee03ae85b15aa0b7a37703e
Former-commit-id: e882ae487e6b1691591fb90e88a3839634ba70fe