Add file completion to tcsh git completion.
For bash completion, the option '-o bashdefault' is used to indicate
that when no other choices are available, file completion should be
performed.  Since this option is not available in tcsh, no file
completion is ever performed.  Therefore, commands like 'git add ',
'git send-email ', etc, require the user to manually type out
the file name.  This can be quite annoying.

To improve the user experience we try to simulate file completion
directly in this script (although not perfectly).

The known issues with the file completion simulation are:
- Possible completions are shown with their directory prefix.
- Completions containing shell variables are not handled.
- Completions with ~ as the first character are not handled.

Signed-off-by: Marc Khouzam <marc.khouzam@ericsson.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2a2828eb21168bf2679142982edad2ef09ab9c31 [formerly 2a2828eb21168bf2679142982edad2ef09ab9c31 [formerly 75ed918bda00a0786bab7de3d220765226fe3f1e]]
Former-commit-id: 2117fbbb6f6e51678f8ba0370b8f6f5bb531e16a
Former-commit-id: 64558e6925186f91e0841d080012525e2fd9defe