git-prompt.sh: update PROMPT_COMMAND documentation
The description of __git_ps1 function operating in two-arg mode was
not very clear.  It said "set PROMPT_COMMAND=__git_ps1" which is not
the right usage for this mode, followed by "To customize the prompt,
do this", giving a false impression that those who do not want to
customize it can get away with no-arg form, which was incorrect.

Make it clear that this mode always takes two arguments, pre and
post, with an example.

The straight-forward one should be listed as the primary usage, and
the confusing one should be an alternate for advanced users.  Swap
the order of these two.

Acked-by: Simon Oosthoek <s.oosthoek@xs4all.nl>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fee0727dd6b14e0b2c3ffc51094e2ec96b7da95a [formerly fee0727dd6b14e0b2c3ffc51094e2ec96b7da95a [formerly de29a7ac0ec31d8f5fa91b7d2f0294165787d558]]
Former-commit-id: 150c82176f7e628b8fce0815d81e918df1b371e4
Former-commit-id: ce89db1fa190dcff0fb96d99d987c50e3fd00687