mailmap: clean up read_mailmap error handling
The error handling for the read_mailmap function is odd. It
returns 1 on error, rather than -1. And it treats a
non-existent mailmap as an error, even though there is no
reason that one needs to exist. Unless some other mailmap
source loads successfully, in which case the original error
is completely masked.

This does not cause any bugs, however, because no caller
bothers to check the return value, anyway. Let's make this a
little more robust to real errors and less surprising for
future callers that do check the error code:

  1. Return -1 on errors.

  2. Treat a missing entry (e.g., no mailmap.file given),
     ENOENT, or a non-existent blob (for mailmap.blob) as
     "no error".

  3. Complain loudly when a real error (e.g., a transient
     I/O error, no permission to open the mailmap file,
     missing or corrupted blob object, etc) occurs.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 095d123358d681b12cdaab78b5e51ac1906d4f4d [formerly 095d123358d681b12cdaab78b5e51ac1906d4f4d [formerly 938a60d64fc91fd7d646ed33fad527b724d1e534]]
Former-commit-id: 215b785ce48510d9e2c7ff2e42c349e098ea0f44
Former-commit-id: 2b23540eda114c7343be01bbc53ff9d905138cd3