index-format.txt: clarify what is "invalid"
A cache-tree entry with a negative entry count is considered invalid
by the current Git; it records that we do not know the object name
of a tree that would result by writing the directory covered by the
cache-tree as a tree object.

Clarify that any entry with a negative entry count is invalid, but
the implementations must write -1 there. This way, we can later
decide to allow writers to use negative values other than -1 to
encode optional information on such invalidated entries without
harming interoperability; we do not know what will be encoded and
how, so we keep these other negative values as reserved for now.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 45f0e3d65da6c1dced34112a696942a44b7c6397 [formerly 45f0e3d65da6c1dced34112a696942a44b7c6397 [formerly 4a6385fe555f810e8168399df673d00addffb721]]
Former-commit-id: f9a770fe79c42bd0ba150418e78bdc30d62087d0
Former-commit-id: 221b41411b89631a4c6c74cdd29b4933317490d0